SoilR-package {SoilR}	R Documentation

SoilR

Description

The package allows you to study compartmental Soil models.

Details

The typical workflow consists of the following steps:

Create a model(run)

Inspect it

The simplest way of creating a model is to use one of the top level functions for predefined models: predefinedModels. The objects returned by these functions can be of different type, usually either

Model

Model_14 .

To inspect the behavior of a model object these classes provide several methods to be found in their respective descriptions. If none of the predefined models fits your needs you can assemble your own model. The functions that create it are the constructors of the above mentioned classes. By convention they have the same name as the class and are described here:

Model

Model_14 .

[Package SoilR version 1.2-98 Index]