AbsoluteFractionModern_from_Delta14C,numeric-method {SoilR}	R Documentation

AbsoluteFractionModern_from_Delta14C,numeric-method Converts from Delta14C to Absolute Fraction Modern

Description

Converts a number or vector containing Delta14C values to the appropriate Absolute Fraction Modern values. Have a look at the methods for details.

Usage

## S4 method for signature 'numeric'
AbsoluteFractionModern_from_Delta14C(delta14C)

Arguments

delta14C
: of class numeric, A numeric object containing the values in Delta14C format

[Package SoilR version 1.1-40 Index]