AbsoluteFractionModern_from_Delta14C,matrix-method {SoilR}	R Documentation

AbsoluteFractionModern_from_Delta14C,matrix-method Converts from Delta14C to Absolute Fraction Modern

Description

This method produces a matrix of Delta14C values from a Matrix or number of Absolute Fraction Modern

Usage

## S4 method for signature 'matrix'
AbsoluteFractionModern_from_Delta14C(delta14C)

Arguments

delta14C
: of class matrix, An object of class matrix containing the values in Delta14C format

[Package SoilR version 1.1-48 Index]