BoundFc-class {SoilR}	R Documentation

Objects containing the atmospheric 14C fraction and the format it is provided in.

Description

no Description

Methods

No methods defined with class BoundFc in the signature.

Exported superclasses

TimeMap-class

Constructors

BoundFc
Please also look at constructors of non virtual subclasses

[Package SoilR version 1.1-36 Index]