BoundFc,character-method {SoilR}	R Documentation

BoundFc,character-method A constructor

Description

The method constructs an object from a format string and other parameters which are passed on to TimeMap

Usage

## S4 method for signature 'character'
BoundFc(format,
...)

Arguments

format
: of class character, a string that specifies the format used to represent the atmospheric fraction. Possible values are "Delta14C" which is the default or "afn" the Absolute Fraction Normal representation

...
: passed on to TimeMap

[Package SoilR version 1.1-53 Index]