BoundFc,ANY,missing,missing-method {SoilR}	R Documentation

BoundFc,ANY,missing,missing-method constructor

Description

the method constructs an object from a function a timerange where it is valid and a format

Usage

## S4 method for signature 'ANY,missing,missing'
BoundFc(map,
lag=0,
format,
interpolation=splinefun)

Arguments

map
: of class ANY, anything that can be used as a single map parameter

lag
: a scalar describing the time lag. Positive Values shift the argument of the interpolation function forward in time. (retard its effect)

format
: a string that specifies the format used to represent the atmospheric fraction. Possible values are "Delta14C" which is the default or "afn" the Absolute Fraction Normal representation

interpolation
: the interpolation function to be used

[Package SoilR version 1.1-36 Index]