BoundLinDecompOp,UnBoundLinDecompOp-method {SoilR}	R Documentation

A converter

Description

The destinction between the classes BoundLinDecompOp and UnboundLinDecompOp exist for those functions, that should be only defined for objects of class UnBoundLinDecomp.

Many functions however do not need extra methods for objects of class UnBoundLinDecompOp and just treat it as a BoundLinDecompOp which is defined on the complete timeline (-Inf,+Inf). With its default arguments this function converts its map argument to a BoundLinDecompOp with just this domain. This is the most frequent internal use case. If starttime and endtime are provided the domain of the operator will be restricted [starttime,endtime].

Usage

## S4 method for signature 'UnBoundLinDecompOp'
BoundLinDecompOp(map, starttime = -Inf, endtime = Inf)

Arguments

map
object of class:UnBoundLinDecompOp, An object of class UnBoundLinDecompOp

starttime
Begin of time interval map will be restricted to

endtime
End of time interval map will be restricted to

[Package SoilR version 1.1-93 Index]