BoundLinDecompOp-class {SoilR}	R Documentation

A S4 class to represent a linear compartmental operator defined on time interval

Description

A S4 class to represent a linear compartmental operator defined on time interval

S4-methods

S4-methods with class BoundLinDecompOp in their signature:

getCompartmentalMatrixFunc,BoundLinDecompOp,ANY,ANY-method

S4-methods with superclasses (in the package) of class BoundLinDecompOp in their signature:

superclass DecompOp

GeneralDecompOp,DecompOp-method

superclass TimeMap

add_plot,TimeMap-method

as.character,TimeMap-method

GeneralDecompOp,TimeMap-method

getFunctionDefinition,TimeMap-method

getTimeRange,TimeMap-method

InFluxes,TimeMap-method

initialize,TimeMap-method

plot,TimeMap-method

TimeMap,TimeMap,ANY,ANY,ANY,ANY-method

S4-superclasses (in the package)

DecompOp

TimeMap

[Package SoilR version 1.2.105 Index]