BoundLinDecompOp,ANY,ANY,ANY-method {SoilR}	R Documentation

BoundLinDecompOp,ANY,ANY,ANY-method a constructor

Description

Creates a BoundLinDecompOp Object.

Usage

## S4 method for signature 'ANY,ANY,ANY'
BoundLinDecompOp(map,
starttime,
endtime,
lag=0,
interpolation=splinefun)

Arguments

map
: of class ANY, anything that can be used as the map parameter

starttime
: of class ANY, for TimeMap

endtime
: of class ANY, for TimeMap

lag
: a scalar describing the time lag. Positive Values shift the argument of the interpolation function forward in time. (retard its effect)

interpolation
: the interpolation function to be used

[Package SoilR version 1.1-36 Index]