BoundLinDecompOp,ANY,missing,missing-method {SoilR}	R Documentation

BoundLinDecompOp,ANY,missing,missing-method a constructor

Description

Creates a BoundLinDecompOp Object.

Usage

## S4 method for signature 'ANY,missing,missing'
BoundLinDecompOp(map,
lag=0,
interpolation=splinefun)

Arguments

map
: of class ANY, anything that can be used as a single map parameter

lag
: a scalar describing the time lag. Positive Values shift the argument of the interpolation function forward in time. (retard its effect)

interpolation
: the interpolation function to be used

[Package SoilR version 1.1-36 Index]