BoundLinDecompOp,function,numeric,numeric-method {SoilR}	R Documentation

BoundLinDecompOp,function,numeric,numeric-method

Description

This method creates a BoundLinDecompOp from a timedependent function and its domain

Usage

## S4 method for signature 'function,numeric,numeric'
BoundLinDecompOp(map,
starttime,
endtime)

Arguments

map
: of class function, a function

starttime
: of class numeric, the begin of the time domain

endtime
: of class numeric, the end of the time domain

[Package SoilR version 1.1-36 Index]