BoundLinDecompOp,UnBoundLinDecompOp-method {SoilR}	R Documentation

BoundLinDecompOp,UnBoundLinDecompOp-method convert a UnBoundLinDecompOp to a BoundLinDecompOp

Description

The method creates a BoundLinDecompOp consisting of a constant time dependent function and the limits of its domain (starttime and endtime) set to -Inf and Inf respectively

Usage

## S4 method for signature 'UnBoundLinDecompOp'
BoundLinDecompOp(map,
starttime=-Inf,
endtime=Inf)

Arguments

map
: of class UnBoundLinDecompOp

starttime
: the left hand boundary of the valid time interval

endtime
: the right hand boundary of the valid time interval

[Package SoilR version 1.1-46 Index]