ConstInFluxes-class {SoilR}	R Documentation

S4 class for a constant influx vector

Description

It is mainly used to dispatch S4-methods for computations that are valid only if the influx is constant. This knowledge can either be used to speed up computations or to decide if they are possible at all. E.g. the computation of equilibria for a model run requires autonomy of the model which requires the influxes to be time independent. If the model is linear and compartmental then the (unique) equilibrium can be computed. Accordingly a method with ConstInFluxes in the signature can be implemented, whereas none would be available for a general InFluxes argument.

S4-methods

S4-methods with class ConstInFluxes in their signature:

ConstantInFluxList_by_PoolIndex,ConstInFluxes-method

getConstantInFluxVector,ConstInFluxes-method

getFunctionDefinition,ConstInFluxes-method

getTimeRange,ConstInFluxes-method

S4-methods with superclasses (in the package) of class ConstInFluxes in their signature:

superclass InFluxes

InFluxes,InFluxes-method

[Package SoilR version 1.1-93 Index]