ConstLinDecompOp {SoilR}	R Documentation

Generic constructor for the class with the same name

Description

Generic constructor for the class with the same name

Usage

ConstLinDecompOp(
  mat,
  internal_flux_rates,
  out_flux_rates,
  numberOfPools,
  poolNames
)

Arguments

mat
see method arguments

internal_flux_rates
see method arguments

out_flux_rates
see method arguments

numberOfPools
see method arguments

poolNames
see method arguments

S4-methods

ConstLinDecompOp,matrix,missing,missing,missing,missing-method

ConstLinDecompOp,missing,ConstantInternalFluxRateList_by_PoolIndex,ConstantOutFluxRateList_by_PoolIndex,numeric,missing-method

ConstLinDecompOp,missing,ConstantInternalFluxRateList_by_PoolIndex,missing,numeric,missing-method

ConstLinDecompOp,missing,ConstantInternalFluxRateList_by_PoolName,ConstantOutFluxRateList_by_PoolName,missing,character-method

ConstLinDecompOp,missing,missing,ConstantOutFluxRateList_by_PoolIndex,numeric,missing-method

[Package SoilR version 1.2.105 Index]