ConstLinDecompOp-class {SoilR}	R Documentation

A class to represent a constant (=nonautonomous,linear) compartmental matrix or equivalently a combination of ordered constant internal flux rates and constant out flux rates.

Description

A class to represent a constant (=nonautonomous,linear) compartmental matrix or equivalently a combination of ordered constant internal flux rates and constant out flux rates.

S4-methods

S4-methods with class ConstLinDecompOp in their signature:

getCompartmentalMatrixFunc,ConstLinDecompOp,ANY,ANY-method

getConstantCompartmentalMatrix,ConstLinDecompOp-method

getConstantInternalFluxRateList_by_PoolIndex,ConstLinDecompOp-method

getConstantOutFluxRateList_by_PoolIndex,ConstLinDecompOp-method

getFunctionDefinition,ConstLinDecompOp-method

getMeanTransitTime,ConstLinDecompOp-method

getTimeRange,ConstLinDecompOp-method

getTransitTimeDistributionDensity,ConstLinDecompOp-method

initialize,ConstLinDecompOp-method

S4-methods with superclasses (in the package) of class ConstLinDecompOp in their signature:

superclass DecompOp

GeneralDecompOp,DecompOp-method

[Package SoilR version 1.1-97 Index]