ConstLinDecompOp,matrix-method {SoilR}	R Documentation

ConstLinDecompOp,matrix-method construct from matric

Description

This method creates a ConstLinDecompOp from a matrix The operator is assumed to act on the vector of carbon stocks by multiplication of the (time invariant) matrix from the left.

Usage

## S4 method for signature 'matrix'
ConstLinDecompOp(mat)

Arguments

mat
: of class matrix

[Package SoilR version 1.1-36 Index]