ConstLinDecompOpWithLinearScalarFactor-class {SoilR}	R Documentation

A class to represent a constant (=nonautonomous,linear) compartmental matrix with a time dependent (linear) scalar pre factor This is a special case of a linear compartmental operator/matrix

Description

A class to represent a constant (=nonautonomous,linear) compartmental matrix with a time dependent (linear) scalar pre factor This is a special case of a linear compartmental operator/matrix

S4-methods

S4-methods with class ConstLinDecompOpWithLinearScalarFactor in their signature:

getConstantCompartmentalMatrix,ConstLinDecompOpWithLinearScalarFactor-method

getConstLinDecompOp,ConstLinDecompOpWithLinearScalarFactor-method

getFunctionDefinition,ConstLinDecompOpWithLinearScalarFactor-method

getLinearScaleFactor,ConstLinDecompOpWithLinearScalarFactor-method

getTimeRange,ConstLinDecompOpWithLinearScalarFactor-method

S4-methods with superclasses (in the package) of class ConstLinDecompOpWithLinearScalarFactor in their signature:

superclass DecompOp

GeneralDecompOp,DecompOp-method

[Package SoilR version 1.2.104 Index]