ConstLinDecompOpWithLinearScalarFactor,matrix,missing,missing,missing,ScalarTimeMap-method {SoilR}	R Documentation

convert names of vectors or lists to class ConstantOutFluxRate convert names of vectors or lists to class ConstantInternalFluxRate

Description

convert names of vectors or lists to class ConstantOutFluxRate convert names of vectors or lists to class ConstantInternalFluxRate

Usage

## S4 method for signature 'matrix,missing,missing,missing,ScalarTimeMap'
ConstLinDecompOpWithLinearScalarFactor(mat, xi)

Arguments

mat
object of class:matrix, no manual documentation

xi
object of class:ScalarTimeMap, no manual documentation

[Package SoilR version 1.2-98 Index]