ConstantInternalFluxRateList_by_PoolIndex-class {SoilR}	R Documentation

Describes a list of flux rates.

Description

The purpose is to avoid creation of lists that contain negative rates or in accidental confusion with list of fluxes. Instances are usually automatically created from data

S4-methods

S4-methods with class ConstantInternalFluxRateList_by_PoolIndex in their signature:

by_PoolName,ConstantInternalFluxRateList_by_PoolIndex-method

ConstLinDecompOp,missing,ConstantInternalFluxRateList_by_PoolIndex,ConstantOutFluxRateList_by_PoolIndex,numeric,missing-method

ConstLinDecompOp,missing,ConstantInternalFluxRateList_by_PoolIndex,missing,numeric,missing-method

[Package SoilR version 1.2.105 Index]