ConstantInternalFluxRateList_by_PoolIndex,list-method {SoilR}	R Documentation

constructor from a normal list

Description

constructor from a normal list

Usage

## S4 method for signature 'list'
ConstantInternalFluxRateList_by_PoolIndex(object)

Arguments

object
object of class:list, A list. Either a list of elements of type ConstantInternalFluxRate_by_PoolIndex or a list where the names of the elements are strings of the form '1->3' (for the flux rate from pool 1 to 2)

Value

An object of class ConstantInternalFluxRateList_by_PoolIndex

The function checks if the elements are of the desired type or can be converted to it. It is mainly used internally and usually called by the front end functions to convert the user supplied arguments.

[Package SoilR version 1.2.105 Index]