ConstantInternalFluxRate_by_PoolIndex-class {SoilR}	R Documentation

S4 class representing a constant internal flux rate

Description

The class is used to dispatch specific methods for the creation of the compartmental matrix which is simplified in case of constant rates.

S4-methods

S4-methods with class ConstantInternalFluxRate_by_PoolIndex in their signature:

by_PoolName,ConstantInternalFluxRate_by_PoolIndex-method

[Package SoilR version 1.1-97 Index]