ConstantOutFluxRateList_by_PoolName-class {SoilR}	R Documentation

Describes a list of flux rates.

Description

The purpose is to avoid creation of lists that contain negative rates or in accidental confusion with list of fluxes. Instances are usually automatically created from data

S4-methods

S4-methods with class ConstantOutFluxRateList_by_PoolName in their signature:

by_PoolIndex,ConstantOutFluxRateList_by_PoolName,ANY,ANY-method

ConstLinDecompOp,missing,ConstantInternalFluxRateList_by_PoolName,ConstantOutFluxRateList_by_PoolName,missing,character-method

[Package SoilR version 1.2.105 Index]