ConstantOutFluxRateList_by_PoolName,list-method {SoilR}	R Documentation

constructor from a normal list

Description

constructor from a normal list

Usage

## S4 method for signature 'list'
ConstantOutFluxRateList_by_PoolName(object)

Arguments

object
object of class:list, A list. Either a list of elements of type ConstantOutFluxRate_by_PoolName or a list where the names of the elements are integer strings of the form '3' (for the flux rate from pool 3)

Value

An object of class ConstantOutFluxRateList_by_PoolName

The function checks if the elements are of the desired type or can be converted to it. It is mainly used internally and usually called by the front end functions to convert the user supplied arguments.

[Package SoilR version 1.1-97 Index]