DecompOp-class {SoilR}	R Documentation

decomposition operator

Description

The decomposition operator is a necessary ingredient of any Model. Very generally it describes the fluxes between the pools and to the exterior as functions of time and the pool contents. SoilR arranges different decomposition operators into different classes, to determine which computations can be performed with them (which methods can be called on them). The simplest and least general decomposition operator is ConstLinDecompOp-class which can be created from a constant reservoir matrix. Since it is the least general (most specific) the additional information can be used to compute more and more specific results (more methods) not available for the more abstract sub classes (for instance BoundLinDecompOp-class that are necessary to model the more general situations where the decomposition and transfer rates are functions of time. The different decomposition operators are created by class-specific functions, called constructors. (see the constructors section of this help page)

Methods

Exported methods directly defined for class DecompOp:

GeneralDecompOp

signature(object = "DecompOp"): ...

GeneralDecompOp,DecompOp-method

[Package SoilR version 1.1-82 Index]