Delta14C_from_AbsoluteFractionModern,numeric-method {SoilR}	R Documentation

Delta14C_from_AbsoluteFractionModern,numeric-method

Description

This method produces Delta14C values from Absolute Fraction Modern Have a look at the methods for details.

Usage

## S4 method for signature 'numeric'
Delta14C_from_AbsoluteFractionModern(AbsoluteFractionModern)

Arguments

AbsoluteFractionModern
: of class numeric, A numeric object containing the values in Absolute Fraction Modern format

[Package SoilR version 1.1-36 Index]