Delta14C_from_AbsoluteFractionModern,matrix-method {SoilR}	R Documentation

Delta14C_from_AbsoluteFractionModern,matrix-method Converts Absolute Fraction Modern values to Delta14C

Description

This method produces a matrix of Delta14C values from a matrix of values in Absolute Fraction Modern.

Usage

## S4 method for signature 'matrix'
Delta14C_from_AbsoluteFractionModern(AbsoluteFractionModern)

Arguments

AbsoluteFractionModern
: of class matrix, An object of class matrix containing the values in Absolute Fraction Modern format

[Package SoilR version 1.1-53 Index]