Fc-class {SoilR}	R Documentation

Fc S4 class

Description

The 14C fraction is a necessary ingredient of any Model_14 object. In the most general case it is a real valued function of time, accompanied by a string describing the unit or format (i.e. "Delta14C" or "afn" for Absolute Fraction Modern) . In the most simple case it is constant real number plus format.

Methods

No exported methods directly defined for class Fc:

Subclasses

BoundFc
ConstFc

Constructors

The class is abstract ( contains "VIRTUAL"). It can therefore not be instanciated directly. Look at non virtual subclasses and their constructors!

[Package SoilR version 1.1-51 Index]