FcAtm.from.Dataframe {SoilR}	R Documentation

FcAtm.from.Dataframe

Description

This function is deprecated constructor of the deprecatied class FcAtm

Usage

FcAtm.from.Dataframe(dframe,
lag=0,
interpolation=splinefun,
format)

Arguments

dframe
A data frame containing exactly two columns: the first one is interpreted as time the secon one is interpreted as atmospheric C14 fraction in the format mentioned

lag
a scalar describing the time lag. Positive Values shift the argument of the interpolation function forward in time. (retard its effect)

interpolation
A function that returns a function the default is splinefun. Other possible values are the linear interpolation approxfun or any self made function with the same interface.

format
a string that specifies the format used to represent the atmospheric fracton. Possible values are "Delta14C" which is the default or "afn" the Absolute Fraction Normal representation

Value

An object of the new class BoundFc that replaces FcAtm

[Package SoilR version 1.1-51 Index]