GeneralDecompOp,DecompOp-method {SoilR}	R Documentation

Pass through factory for objects of subclasses of DecompOp

Description

This method takes and returns an (identical) object that inherits from DecompOp. It's purpose it to be able to call the generic function on arguments that are already

Usage

## S4 method for signature 'DecompOp'
GeneralDecompOp(object)

Arguments

object
object of class:DecompOp, An object that already is of class DecompOp

[Package SoilR version 1.2.105 Index]