GeneralDecompOp,DecompOp-method {SoilR}	R Documentation

GeneralDecompOp,DecompOp-method pass through factory

Description

This method handles the case that no actual construction is necessary since the argument is already of a subclass of DecompOp. See the subclasses section of DecompOp-class

Usage

## S4 method for signature 'DecompOp'
GeneralDecompOp(object)

Arguments

object
: of class DecompOp

[Package SoilR version 1.1-48 Index]