GeneralInFlux,numeric-method {SoilR}	R Documentation

GeneralInFlux,numeric-method conversion of a vector to an object of class ConstInFlux

Description

This method enables the model creating functions to handle constant input streams

Usage

## S4 method for signature 'numeric'
GeneralInFlux(object)

Arguments

object
: of class numeric

[Package SoilR version 1.1-46 Index]