GeneralInFlux,function-method {SoilR}	R Documentation

GeneralInFlux,function-method creates a UnBoundInFlux from a vector valued function

Description

The resulting operator is created by a call to the constructor of class UnBoundInFlux. You should only use this if the domain of your function is the complete time axis (-Inf,+Inf). If your function has a finite domain create an object of class BoundInFlux-class ### by calling BoundInFlux. This will activeate checks on that avoid unintended extrapolation.

Usage

## S4 method for signature 'function'
GeneralInFlux(object)

Arguments

object
: of class function

[Package SoilR version 1.1-46 Index]