GeneralModel_14,numeric,ANY,numeric,ANY,ANY,ANY,missing,numeric,missing,function,logical-method {SoilR}	R Documentation

GeneralModel_14,numeric,ANY,numeric,ANY,ANY,ANY,missing,numeric,missing,function,logical-method a function to create objects of class Model_14

Description

This method tries to create a Model object from any combination of arguments that can be converted into the required set of building blocks for a model for n arbitrarily connected pools.

Usage

## S4 method for signature 
## 'numeric,
##   ANY,
##   numeric,
##   ANY,
##   ANY,
##   ANY,
##   missing,
##   numeric,
##   missing,
##   function,
##   logical'
GeneralModel_14(t,
A,
ivList,
initialValF,
inputFluxes,
inputFc,
di=-0.0001209681,
solverfunc=deSolve.lsoda.wrapper,
pass=FALSE)

Arguments

t
: of class numeric, A vector containing the points in time where the solution is sought.

A
: of class ANY

ivList
: of class numeric

initialValF
: of class ANY

inputFluxes
: of class ANY

inputFc
: of class ANY

di
: of class numeric

solverfunc
: of class function

pass
: of class logical

See Also

TwopParallelModel, TwopSeriesModel, TwopFeedbackModel

[Package SoilR version 1.1-40 Index]