ICBM_N {SoilR}	R Documentation

Implementation of the ICBM/N model

Description

This implementations follows the description in Katterer and Andren (2001, Eco Mod 136:191).

Usage

ICBM_N(
  i = 0.47,
  k_Y = 0.259,
  k_O = 0.0154,
  r_e = 1,
  e_Y = 0.362,
  h = 0.243,
  q_i = 18.8,
  q_b = 5
)

Arguments

i
carbon input to the soil from plant production

k_Y
decomposition rate of young pool Y

k_O
decomposition rate of old pool O

r_e
external effects on decomposition rates

e_Y
yield efficiency of the soil organism community

h
humification coefficient. Fraction of outflux from Y that is not respired and enters O

q_i
C:N ratio of plant inputs

q_b
C:N ratio of soil organism biomass

[Package SoilR version 1.2.103 Index]