InFlux_by_PoolName {SoilR}	R Documentation

Generic constructor for an influx to a single pool from an ordered pair of PoolName (string like) and function objects

Description

Generic constructor for an influx to a single pool from an ordered pair of PoolName (string like) and function objects

Usage

InFlux_by_PoolName(func, destinationName)

Arguments

func
see method arguments

destinationName
see method arguments

[Package SoilR version 1.1-93 Index]