InFlux_by_PoolName,function,character-method {SoilR}	R Documentation

Constructor from an ordered pair of PoolName (string like) and function objects

Description

Constructor from an ordered pair of PoolName (string like) and function objects

Usage

## S4 method for signature ''function',character'
InFlux_by_PoolName(func, destinationName)

Arguments

func
object of class:function, A function. The names of the formal arguments have to be a subset of the state variable names and the time symbol This allows subsequent automatic reordering of the state variables. In the presence of a vector of state-variable-names the formulation can automatically be transformed to a function of a s tate VECTOR argument and time

destinationName
object of class:character, PoolName (string like) object and a function

[Package SoilR version 1.1-93 Index]