InternalFluxList_by_PoolName-class {SoilR}	R Documentation

S4-class for a list of internal fluxes with indexed by (source and destination pool) names

Description

S4-class for a list of internal fluxes with indexed by (source and destination pool) names

S4-methods

S4-methods with class InternalFluxList_by_PoolName in their signature:

as.numeric,InternalFluxList_by_PoolName-method

by_PoolIndex,InternalFluxList_by_PoolName,character,character-method

UnBoundNonLinDecompOp_by_PoolNames,InternalFluxList_by_PoolName,OutFluxList_by_PoolName,character-method

[Package SoilR version 1.2.105 Index]