InternalFlux_by_PoolIndex,function,numeric,numeric,missing-method {SoilR}	R Documentation

constructor from an ordered pair of PoolIndex (integer like) objects and a function with vector argument

Description

constructor from an ordered pair of PoolIndex (integer like) objects and a function with vector argument

Usage

## S4 method for signature ''function',numeric,numeric,missing'
InternalFlux_by_PoolIndex(func, sourceIndex, destinationIndex)

Arguments

func
object of class:function, A function f(X,t) where X is a vector of the state variables. This form is required internally by the solvers and supported for backward compatibility with earlier versions of SoilR. Note that the function func given in this form can not be transformed to a different ordering of state variables, since the location of a state variable in the vector argument depends on a specific order and will be 'hardcoded' into your function. See InternalFlux_by_PoolName for the new more powerful interface which allows subsequent reordering of the state variables by using the names of the state variables as formal arguments for func. In this case SoilR can infer (and later adapt) the vector argument form needed for the solvers. constructor from an ordered pair of PoolIndex (integer like) objects

sourceIndex
object of class:numeric, no manual documentation

destinationIndex
object of class:numeric, no manual documentation

[Package SoilR version 1.1-97 Index]