InternalFlux_by_PoolName {SoilR}	R Documentation

Generic constructor for the class with the same name

Description

Generic constructor for the class with the same name

Usage

InternalFlux_by_PoolName(func, sourceName, destinationName, src_to_dest)

Arguments

func
see method arguments

sourceName
see method arguments

destinationName
see method arguments

src_to_dest
see method arguments

S4-methods

InternalFlux_by_PoolName,function,character,character,missing-method

InternalFlux_by_PoolName,function,missing,missing,character-method

[Package SoilR version 1.2.105 Index]