InternalFlux_by_PoolName-class {SoilR}	R Documentation

S4-class for a single internal flux with source and destination pools specified by name

Description

S4-class for a single internal flux with source and destination pools specified by name

S4-methods

S4-methods with class InternalFlux_by_PoolName in their signature:

as.numeric,InternalFlux_by_PoolName-method

by_PoolIndex,InternalFlux_by_PoolName,character,character-method

[Package SoilR version 1.2.105 Index]