InternalFlux_by_PoolName,function,character,character,missing-method {SoilR}	R Documentation

constructor from an ordered pair of PoolName (string like) objects and a function with the set of formal argument names forming a subset of the state_variable_names

Description

constructor from an ordered pair of PoolName (string like) objects and a function with the set of formal argument names forming a subset of the state_variable_names

Usage

## S4 method for signature ''function',character,character,missing'
InternalFlux_by_PoolName(func, sourceName, destinationName)

Arguments

func
object of class:function, A real valued function describing the flux (mass/time) as function of the state variables and time.

sourceName
object of class:character, A string identifying the source pool of the flux

destinationName
object of class:character, A string identifying the destination pool of the flux

[Package SoilR version 1.1-97 Index]