Model {SoilR}	R Documentation

Constructor for class Model-class

Description

This function creates an object of class Model-class, The arguments can be given in different form as long as they can be converted to the necessary internal building blocks. (See the links)

Usage

Model(t,
A,
ivList,
inputFluxes,
solverfunc=deSolve.lsoda.wrapper,
pass=FALSE)

Arguments

t
A vector containing the points in time where the solution is sought.

A
something that can be converted by GeneralDecompOp to any of the available subclasses of DecompOp-class.

ivList
A numeric vector containing the initial amount of carbon for the n pools. The length of this vector is equal to the number of pools .This is checked by an internal function.

inputFluxes
something that can be converted by GeneralInFlux to any of the available subclasses of InFlux-class.

solverfunc
The function used to actually solve the ODE system. This can be deSolve.lsoda.wrapper or any other user provided function with the same interface.

pass
Forces the constructor to create the model even if it does not pass internal sanity checks

Details

The internal constructor of class Model-class requires the argument A to be of class DecompOp-class and argument inputFluxes to be of class InFlux-class. Before calling the internal constructor this function calls GeneralDecompOp on its argument A and GeneralInFlux on its argument inputFluxes to convert them into the required classes. Both are generic functions. Follow the links to see for which kind of inputs conversion methods are available. The attempted conversion allows great flexibility with respect to arguments and independence from the actual implementation. However if your code uses the wrong argument the error will most likely occur in the delegate functions. If this happens use traceback() to see which function was called and try to call the constructor of the desired subclass explicitly with your arguments. The subclasses are linked in the class documentation DecompOp-class or InFlux-class respectively.

Note also that this functions checks its arguments quite elaborately and tries to detect accidental unreasonable combinations, especially concerning two kinds of error.

unintended extrapolation of time series data

violoations of massbalance by the DecompOp argument.

SoilR has a lot of unit tests which are installed i with the package and are sometimes instructive as examples. To see example scenarios for parameter check look at: > system.file("tests", "runit.correctness_of_Model.R", package = "SoilR") [1] "/home/mm/R/x86_64-pc-linux-gnu-library/3.4/SoilR/tests/runit.correctness_of_Model.R"

Value

An object of class Model-class that can be queried by many methods to be found there.

See Also

This function is called by many of the predefinedModels

Examples

# examples from external files
# inst/examples/ModelExamples.R CorrectNonautonomousLinearModelExplicit:

  # This example describes the creation and use of a Model object that 
  # is defined by time dependent functions for decomposition and influx.
  # The constructor of the Model-class  (see  ?Model) is a 'Generic' 
  # function  that  has 'Methods' for different combinations of 
  # arguments.
  # In this example we will call the constructor whith arguments which 
  # are of the same type as the current internal representations in the 
  # Model object. and create these arguments explicitly beforehand. 
  # See other examples for the generic constructor Model 
  # We start with the Decomposition Operator.
  # For this example we assume that we are able to describe it 
  # by explicit R functions and therefore choose the appropriate
  # sub class BoundLinDecompOp
  # of DecompOp explicitly.  (see ?'BoundLinDecompOp-class') 
  A=BoundLinDecompOp(
    ## We call the generic constructor (see ?BoundLindDcompOp) 
    ## which has a method  
    ## that takes a matrix-valued function of time as its first argument.
    function(t){
      matrix(nrow=3,ncol=3,byrow=TRUE,
         c(
           -1,    0,        0,
          0.5,   -2,        0,
            0,    1, sin(t)-1 
        )
      )    
    },
    ## The other two arguments describe the time interval where the 
    ## function is valid (the domain of the function)
    ## The interval will be checked against the domain of the InFlux
    ## argument of Model and against its 't' argument to avoid 
    ## invalid computations outside the domain. 
    ##  Inf and -Inf are possible values. 
    starttime=0,
    endtime=20
  )  
  I=BoundInFlux(
     ## The first argument is a vector-valued function of time
     function(t){
       matrix(nrow=3,ncol=1,byrow=TRUE,
           c(-1,    0,    0)
       )
     },
     ## The other two arguments describe the time interval where the 
     ## function is valid (the domain of the function)
     starttime=0,
     endtime=40
  )
  ## No we specify the points in time where we want 
  ## to compute results
  t_start=0 
  t_end=10 
  tn=50
  timestep <- (t_end-t_start)/tn 
  times <- seq(t_start,t_end,timestep) 
  ## and the start values
  sv=c(0,0,0)
  mod=Model(t=times,A,sv,I)

  ## No we use the model to compute some results
  getC(mod)
  getReleaseFlux(mod)
  #also look at the methods section of Model-class 

[Package SoilR version 1.1-36 Index]