Model_14-class {SoilR}	R Documentation

Model_14

Description

This class extends Model, to represent 14C decay.

As Model it contains all the components that are needed to solve the initial value problem for the pool contents C=(C_1,...C_n)^t.

It adds the components that are needed to solve the additional initial value problem for the 14C contents of the pools 14C=(14C_1,...,14C_n)^t.

It provides the single argument for all the functions that are available for an argument of class Model.

and for additional functions that are available to compute various results from the solution of the additional initial value problem for 14C. See subsection Methods and the examples.)

Details

The original initial value problem for (C_1,...,C_n)^t) as decribed in the docomentation of the superclass Model was given by:

dC(t)/dt = I(t) + A(t)C(t)

the initial Values C_0=C(t_0)

for the times {t_0,....,t_m}.

The additional initial value problem for 14C is represented by additional parameters:

a second ordinary differential equation:

d 14C/dt = F I(t) + A(t)C(t)-k_14 C(t)

which is related to the original system by 14C=F C resulting in :

d 14C/dt = F ((A(t)C(t)-k_14) C(t)+I(t)

the 14C fraction F

the 14C decay rate k_14.

In an object of class Model_14 the components are represented as follows:

The time-dependent matrix valued function A(t) is represented by an object of a subclass of DecompOp (for decomposition operator). Such objects can be created in different ways from functions, matrices or data. (see the subclasses of DecompOp and especially their Constructors sections. and the examples section of this help page.

The vector-valued time-dependent function I(t) is in SoilR represented by an object of a subclass of class InFlux. Such objects can also be created from functions, constant vectors and data. (see the subclasses of InFlux and especially their Constructors sections.

The initial values for C are represented by a numeric vector (Note that the initial values for 14C are given by 14C_0=F C_0.)

An object of a subclass of Fc representing the 14C fraction F and its unit. (Either "Delta14C" or "afn" for Absolute Fraction Normal)

Methods

getC14

signature(object = "Model_14"): ...

getC14,Model_14-method

getF14

signature(object = "Model_14"): ...

getF14,Model_14-method

getF14C

signature(object = "Model_14"): ...

getF14C,Model_14-method

getF14R

signature(object = "Model_14"): ...

getF14R,Model_14-method

getReleaseFlux14

signature(object = "Model_14"): ...

getReleaseFlux14,Model_14-method

Superclasses

Model-class

Constructors

Model_14

[Package SoilR version 1.1-50 Index]