Model_by_PoolNames {SoilR}	R Documentation

Constructor for Model_by_PoolNames

Description

Constructor for Model_by_PoolNames

Usage

Model_by_PoolNames(
  smod,
  times,
  mat,
  initialValues,
  inputFluxes,
  internal_fluxes,
  out_fluxes,
  timeSymbol,
  solverfunc
)

Arguments

smod
see method arguments

times
see method arguments

mat
see method arguments

initialValues
see method arguments

inputFluxes
see method arguments

internal_fluxes
see method arguments

out_fluxes
see method arguments

timeSymbol
see method arguments

solverfunc
see method arguments

Value

A possibly nonlinear Model(run) that contains information about the pool names and connectivity of the pools and is therefore the preferred representation for new code.

S4-methods

Model_by_PoolNames,missing,numeric,UnBoundNonLinDecompOp_by_PoolNames,numeric,InFluxList_by_PoolName,missing,missing,missing,missing-method

Model_by_PoolNames,SymbolicModel_by_PoolNames,numeric,missing,numeric,missing,missing,missing,missing,missing-method

[Package SoilR version 1.2.105 Index]