OutFlux_by_PoolName,function,character-method {SoilR}	R Documentation

constructor from a PoolName (integer like) object and a function

Description

constructor from a PoolName (integer like) object and a function

Usage

## S4 method for signature ''function',character'
OutFlux_by_PoolName(func, sourceName)

Arguments

func
object of class:function, A function. The names of the formal arguments have to be a subset of the state variable names and the time symbol This allows subsequent automatic reordering of the state variables. In the presence of a vector of stave variable names the formulation can automatically be transformed to a function of a state VECTOR argument and #' time constructor from an ordered pair of PoolName (integer like) objects

sourceName
object of class:character, no manual documentation

[Package SoilR version 1.2.104 Index]