PoolConnection_by_PoolIndex-class {SoilR}	R Documentation

Objects that have a source and a destination described by integer like objects ( of class PoolIndex)

Description

Examples are internal Fluxes and Fluxrates Their 'topologic' part and many related sanity checks are implemented here rather than in every function that uses fluxes or rates The methods are also essential for the translation from (internal) flux lists to the respective parts of compartmental matrices and back

[Package SoilR version 1.2.103 Index]