PoolIndex,PoolIndex-method {SoilR}	R Documentation

pass through constructor fron an object of the same class

Description

This is here to be able to call PoolIndex on a PoolIndex object without having to check before if it is necessary. the unnecessary poolNames argument will be ignored.

Usage

## S4 method for signature 'PoolIndex'
PoolIndex(id, poolNames)

Arguments

id
object of class:PoolIndex, no manual documentation

poolNames
no manual documentation

[Package SoilR version 1.2-98 Index]