PoolName-class {SoilR}	R Documentation

class for pool-name-strings

Description

used to control the creation of PoolName objects which have to be valid R identifiers and to dispatch pool name specific methods like conversion to pool indices

S4-methods

S4-methods with class PoolName in their signature:

PoolIndex,PoolName-method

PoolName,PoolName-method

[Package SoilR version 1.2.105 Index]