SoilR.F0.new {SoilR}	R Documentation

deprecated function that used to create an object of class SoilR.F0

Description

The function internally calls the constructor of the replacement class ConstFc-class.

Usage

SoilR.F0.new(values=c(0),
format=Delta14C)

Arguments

values
a numeric vector

format
a character string describing the format e.g. "Delta14C"

Value

An object of class ConstFc-class that contains data and a format description that can later be used to convert the data into other formats if the conversion is implemented.

[Package SoilR version 1.1-51 Index]