TimeMap,TimeMap,ANY,ANY,ANY,ANY-method {SoilR}	R Documentation

TimeMap,TimeMap,ANY,ANY,ANY,ANY-method pass through constructor

Description

The function just returns its argument. So any function that has to convert one of its argument can just call TimeMap on it even if the argument is allready one.

Usage

## S4 method for signature 'TimeMap,ANY,ANY,ANY,ANY'
TimeMap(map)

Arguments

map
: of class TimeMap, the object that will be returned unchanged

[Package SoilR version 1.1-41 Index]