TimeMap,missing,missing,missing,numeric,list-method {SoilR}	R Documentation

TimeMap,missing,missing,missing,numeric,list-method Create a TimeMap from a nested list

Description

The method creates an instance of TimeMap-class from a vector of times and a list of the same length, containing vectors matrices or arrays

Usage

## S4 method for signature 'missing,missing,missing,numeric,list'
TimeMap(times,
data,
lag=0,
interpolation=splinefun)

Arguments

times
: of class numeric, A vector of times

data
: of class list, A list of the same length as times with every list element refering to one time in times

lag
: Either a scalar or an object of the same size as the list elements describing the time lag of the data

interpolation
: the interpolation method to be used

[Package SoilR version 1.1-53 Index]