TimeMap,list,missing,missing,missing,missing-method {SoilR}	R Documentation

TimeMap,list,missing,missing,missing,missing-method Create a TimeMap from a nested list

Description

The method creates an instance of TimeMap-class from a list that contains data and a vector of times referring to it.

Usage

## S4 method for signature 'list,missing,missing,missing,missing'
TimeMap(map)

Arguments

map
: of class list

Details

The list must have two entries If the entries are not named, the first one is supposed to be a numeric vector of times and the second to contain the data referring to those times. The data entry of the list can itself be a list with the same length as the times entry or an array whose last dimension is equal to the length of the times entry. If the data entry is a list the elements must be vectors,matrices or arrays.

[Package SoilR version 1.1-40 Index]