UnBoundNonLinDecompOp_by_PoolNames,InternalFluxList_by_PoolName,OutFluxList_by_PoolName,character-method {SoilR}	R Documentation

A flux and pool name based representation of a possibly nonlinear and nonautonomous Compartmental Matrix

Description

A flux and pool name based representation of a possibly nonlinear and nonautonomous Compartmental Matrix

Usage

## S4 method for signature 
## 'InternalFluxList_by_PoolName,OutFluxList_by_PoolName,character'
UnBoundNonLinDecompOp_by_PoolNames(internal_fluxes, out_fluxes, timeSymbol)

Arguments

internal_fluxes
object of class:InternalFluxList_by_PoolName, InternalFluxList_by_PoolName

out_fluxes
object of class:OutFluxList_by_PoolName, OutFluxList_by_PoolName

timeSymbol
object of class:character,

[Package SoilR version 1.2.105 Index]