UnBoundNonLinDecompOp_by_PoolNames-class {SoilR}	R Documentation

An S4 class to represent the of nonlinear nonautonomous compartmental system independently of the order of state variables

Description

An S4 class to represent the of nonlinear nonautonomous compartmental system independently of the order of state variables

S4-methods

S4-methods with class UnBoundNonLinDecompOp_by_PoolNames in their signature:

getCompartmentalMatrixFunc,UnBoundNonLinDecompOp_by_PoolNames,character,character-method

Model_by_PoolNames,missing,numeric,UnBoundNonLinDecompOp_by_PoolNames,numeric,InFluxList_by_PoolName,missing,missing,missing,missing-method

UnBoundNonLinDecompOp,missing,missing,missing,missing,character,character,UnBoundNonLinDecompOp_by_PoolNames-method

[Package SoilR version 1.2.104 Index]