add_plot,TimeMap-method {SoilR}	R Documentation

add_plot,TimeMap-method overview plot

Description

The method adds a simple overview plot of a the scalar, vector, matrix or arrayvalued function plotting all time dependent component.

Usage

## S4 method for signature 'TimeMap'
add_plot(x,
...)

Arguments

x
: of class TimeMap, the TimeMap Object to be plotted

...
: other plot parameters not implemented yet

[Package SoilR version 1.1-53 Index]