as.character {SoilR}	R Documentation

creates a character representation of the object in question

Description

This function computes the carbon content of the pools as function of time

Usage

as.character(object)

Arguments

object
The object to be printed.

Author(s)

Carlos A. Sierra <csierra@bgc-jena.mpg.de>, Markus Mueller <mamueller@bgc-jena.mpg.de>

[Package SoilR version 1.1-46 Index]