as.numeric,InternalFlux_by_PoolName-method {SoilR}	R Documentation

Convert to a numeric value with name of the form 'a->b'

Description

Convert to a numeric value with name of the form 'a->b'

Usage

## S4 method for signature 'InternalFlux_by_PoolName'
as.numeric(x, y, t, time_symbol, ...)

Arguments

x
object of class:InternalFlux_by_PoolName, The list of fluxes. Every element contains a function that depends on a combination of of state variables and time.

y
A vector indexed by the names of the state variables

t
a number representing the current point in time

time_symbol
The name of the time argument used in the definition of the flux functions

...
no manual documentation

[Package SoilR version 1.1-93 Index]