bind.C14curves {SoilR}	R Documentation

Binding of pre- and post-bomb Delta14C curves

Description

This function takes a pre- and a post-bomb curve, binds them together, and reports the results back either in years BP or AD.

Usage

bind.C14curves(prebomb, postbomb, time.scale)

Arguments

prebomb
A pre-bomb radiocarbon dataset. They could be either IntCal09 or IntCal13.

postbomb
A post-bomb radiocarbon dataset. They could be any of the datasets in Hua2013.

time.scale
A character indicating whether to report the results in years before present BP or anno domini AD.

Value

A data.frame with 3 columns: years in AD or BP, the atmospheric Delta14C value, the standard deviation of the Delta14C value.

[Package SoilR version 1.1-82 Index]