by_PoolIndex,ConstantInFluxRate_by_PoolName,ANY,ANY-method {SoilR}	R Documentation

new object with the source pool id converted to a PoolIndex if necessary

Description

new object with the source pool id converted to a PoolIndex if necessary

new object with the source pool id converted to a PoolIndex if necessary

new object with the source pool id converted to a PoolIndex if necessary

Usage

## S4 method for signature 'ConstantInFluxRate_by_PoolName,ANY,ANY'
by_PoolIndex(obj, poolNames)

## S4 method for signature 'ConstantInFluxRate_by_PoolName,ANY,ANY'
by_PoolIndex(obj, poolNames)

## S4 method for signature 'ConstantInFluxRate_by_PoolName,ANY,ANY'
by_PoolIndex(obj, poolNames)

Arguments

obj
object of class:ConstantInFluxRate_by_PoolName, no manual documentation

poolNames
no manual documentation

[Package SoilR version 1.2.105 Index]