by_PoolIndex,PoolConnection_by_PoolName,ANY,ANY-method {SoilR}	R Documentation

constructor from strings of the form 'x->y' new object with the source pool id and the destination pool id guaranteed to be of class PoolIndex

Description

converts the ids if necessary otherwise returns an identical object

Usage

## S4 method for signature 'PoolConnection_by_PoolName,ANY,ANY'
by_PoolIndex(obj, poolNames)

Arguments

obj
object of class:PoolConnection_by_PoolName, no manual documentation

poolNames
no manual documentation

[Package SoilR version 1.2.104 Index]