by_PoolName,ConstantOutFluxRate_by_PoolIndex-method {SoilR}	R Documentation

new object with the source pool id converted to a PoolName if necessary

Description

This method exists only for classes that do not contain functions of the state_variables since we cannot automatically translate functions with a state vector arguments to functions of the respective state variables which would require symbolic computations. The reverse direction is always possible and is therefore the preferred way to input rate functions that depend on state variables.

Usage

## S4 method for signature 'ConstantOutFluxRate_by_PoolIndex'
by_PoolName(obj, poolNames)

Arguments

obj
object of class:ConstantOutFluxRate_by_PoolIndex, no manual documentation

poolNames
no manual documentation

[Package SoilR version 1.1-93 Index]