check_id_length {SoilR}	R Documentation

helper function to check that the length of the argument is exactly 1

Description

helper function to check that the length of the argument is exactly 1

Usage

check_id_length(id)

Arguments

id
Either a string or a number

[Package SoilR version 1.2.103 Index]