cycling {SoilR}	R Documentation

Cycling analysis of compartmental matrices

Description

Computes the fundamental matrix N, and the expected number of steps from a compartmental matrix A

Usage

cycling(A= )

Arguments

A
A compartmental linear square matrix with cycling rates in the diagonal and transfer rates in the off-diagonal.

Value

A list with 2 objects: the fundamental matrix N, and the expected number of steps Et.

See Also

systemAge

[Package SoilR version 1.1-49 Index]