deSolve.lsoda.wrapper {SoilR}	R Documentation

deSolve.lsoda.wrapper

Description

The function serves as a wrapper for lsoda using a much simpler interface which allows the use of matrices in the definition of the derivative. To use lsoda we have to convert our vectors to lists, define tolerances and so on. This function does this for us , so we don't need to bother about it.

Usage

deSolve.lsoda.wrapper(t,
ydot,
startValues)

Arguments

t
A row vector containing the points in time where the solution is sought.

ydot
The function of y and t that computes the derivative for a given point in time and a column vector y.

startValues
A column vector with the starting values.

Value

A matrix. Every column represents a pool and every row a point in time

[Package SoilR version 1.1-46 Index]