entropyRatePerTime {SoilR}	R Documentation

Entropy rate per time

Description

Computes the entropy rate per time of the Markov chain generated by the compartmental system

Usage

entropyRatePerTime(A, u)

Arguments

A
A constant compartmental square matrix with cycling rates in the diagonal and transfer rates in the off-diagonal.

u
A one-column matrix defining the amount of inputs per compartment.

Value

A scalar value with the entropy rate per time

References

Metzler, H. (2020). Compartmental systems as Markov chains : age, transit time, and entropy (T. Oertel-Jaeger, I. Pavlyukevich, and C. Sierra, Eds.) [PhD thesis](https://suche.thulb.uni-jena.de/Record/1726091651)

Examples

B6=matrix(c(-1,1,0,0,-1,1,0,0,-1),3,3); u6=matrix(c(1,0,0))
entropyRatePerTime(A=B6, u=u6)

[Package SoilR version 1.2.105 Index]