fT.Arrhenius {SoilR}	R Documentation

Effects of temperature on decomposition rates according the Arrhenius equation

Description

Calculates the effects of temperature on decomposition rates according to the Arrhenius equation.

Usage

fT.Arrhenius(Temp,
A=1000,
Ea=75000,
Re=8.3144621)

Arguments

Temp
A scalar or vector containing values of temperature (in degrees Kelvin) for which the effects on decomposition rates are calculated.

A
A scalar defining the pre-exponential factor.

Ea
A scalar defining the activation energy in units of J mol^-1.

Re
A scalar defining the universal gas contant in units of J K^-1 mol^-1.

Value

A scalar or a vector containing the effects of temperature on decomposition rates (unitless).

[Package SoilR version 1.1-40 Index]