fT.Daycent2 {SoilR}	R Documentation

Effects of temperature on decomposition rates according to the DAYCENT model

Description

Calculates the effects of temperature on decomposition rates according to the Daycent/Century models.

Usage

fT.Daycent2(Temp)

Arguments

Temp
A scalar or vector containing values of soil temperature for which the effects on decomposition rates are calculated.

Value

A scalar or a vector containing the effects of temperature on decomposition rates (unitless).

References

Del Grosso, S. J., W. J. Parton, A. R. Mosier, E. A. Holland, E. Pendall, D. S. Schimel, and D. S. Ojima (2005), Modeling soil CO2 emissions from ecosystems, Biogeochemistry, 73(1), 71-91.

[Package SoilR version 1.1-48 Index]