fT.Demeter {SoilR}	R Documentation

Effects of temperature on decomposition rates according to the DEMETER model

Description

Calculates the effects of temperature on decomposition rates according to the DEMETER model.

Usage

fT.Demeter(Temp,
Q10=2)

Arguments

Temp
A scalar or vector containing values of temperature for which the effects on decomposition rates are calculated

Q10
A scalar. Temperature coefficient Q10

Value

A scalar or a vector containing the effects of temperature on decomposition rates (unitless).

References

Foley, J. A. (1995), An equilibrium model of the terrestrial carbon budget, Tellus B, 47(3), 310-319.

[Package SoilR version 1.1-50 Index]