fT.KB {SoilR}	R Documentation

Effects of temperature on decomposition rates according to a model proposed by M. Kirschbaum (1995)

Description

Calculates the effects of temperature on decomposition rates according to a model proposed by Kirschbaum (1995).

Usage

fT.KB(Temp)

Arguments

Temp
a scalar or vector containing values of soil temperature for which the effects on decomposition rates are calculated

Value

A scalar or a vector containing the effects of temperature on decomposition rates (unitless).

References

Kirschbaum, M. U. F. (1995), The temperature dependence of soil organic matter decomposition, and the effect of global warming on soil organic C storage, Soil Biology and Biochemistry, 27(6), 753-760.

[Package SoilR version 1.1-47 Index]