fT.Q10 {SoilR}	R Documentation

Effects of temperature on decomposition rates according to a Q10 function

Description

Calculates the effects of temperature on decomposition rates according to the modified Van't Hoff function (Q10 function).

Usage

fT.Q10(Temp,
k_ref=1,
T_ref=10,
Q10=2)

Arguments

Temp
A scalar or vector containing values of temperature for which the effects on decomposition rates are calculated.

k_ref
A scalar representing the value of the decomposition rate at a reference temperature vaule.

T_ref
A scalar representing the reference temperature.

Q10
A scalar. Temperature coefficient Q10.

Value

A scalar or a vector containing the effects of temperature on decomposition rates (unitless).

[Package SoilR version 1.1-53 Index]