fT.RothC {SoilR}	R Documentation

Effects of temperature on decomposition rates according to the functions included in the RothC model

Description

Calculates the effects of temperature on decomposition rates according to the functions included in the RothC model.

Usage

fT.RothC(Temp)

Arguments

Temp
A scalar or vector containing values of temperature for which the effects on decomposition rates are calculated.

Value

A scalar or a vector containing the effects of temperature on decomposition rates (unitless).

Note

This function returns NA for Temp <= -18.3

References

Jenkinson, D. S., S. P. S. Andrew, J. M. Lynch, M. J. Goss, and P. B. Tinker (1990), The Turnover of Organic Carbon and Nitrogen in Soil, Philosophical Transactions: Biological Sciences, 329(1255), 361-368.

[Package SoilR version 1.1-49 Index]