fT.Standcarb {SoilR}	R Documentation

Effects of temperature on decomposition rates according to the StandCarb model

Description

Calculates the effects of temperature on decomposition rates according to the StandCarb model.

Usage

fT.Standcarb(Temp,
Topt=45,
Tlag=4,
Tshape=15,
Q10=2)

Arguments

Temp
A scalar or vector containing values of temperature for which the effects on decomposition rates are calculated.

Topt
A scalar representing the optimum temperature for decomposition.

Tlag
A scalar that determines the lag of the response curve.

Tshape
A scalar that determines the shape of the response curve.

Q10
A scalar. Temperature coefficient Q10.

Value

A scalar or a vector containing the effects of temperature on decomposition rates (unitless).

References

Harmon, M. E., and J. B. Domingo (2001), A users guide to STANDCARB version 2.0: A model to simulate carbon stores in forest stands. Oregon State University, Corvallis.

[Package SoilR version 1.1-50 Index]