fW.Candy {SoilR}	R Documentation

Effects of moisture on decomposition rates according to the Candy model

Description

Calculates the effects of water content and pore volume on decomposition rates.

Usage

fW.Candy(theta, PV)

Arguments

theta
A scalar or vector containing values of volumetric soil water content.

PV
A scalar or vector containing values of pore volume.

References

J. Bauer, M. Herbst, J.A. Huisman, L. Weiherm\"uller, H. Vereecken. 2008. Sensitivity of simulated soil heterotrophic respiration to temperature and moisture reduction functions. Geoderma, Volume 145, Issues 1-2, 15 May 2008, Pages 17-27.

[Package SoilR version 1.2.104 Index]