fW.Century {SoilR}	R Documentation

Effects of moisture on decomposition rates according to the CENTURY model

Description

Calculates the effects of precipitation and potential evapotranspiration on decomposition rates.

Usage

fW.Century(PPT, PET)

Arguments

PPT
A scalar or vector containing values of monthly precipitation.

PET
A scalar or vector containing values of potential evapotranspiration.

Value

A scalar or a vector containing the effects of precipitation and potential evapotranspiration on decomposition rates (unitless).

[Package SoilR version 1.1-82 Index]