fW.Daycent2 {SoilR}	R Documentation

Effects of moisture on decomposition rates according to the DAYCENT model

Description

Calculates the effects of volumetric water content on decomposition rates according to the Daycent/Century models.

Usage

fW.Daycent2(W,
WP=0,
FC=100)

Arguments

W
A scalar or vector of volumetric water content in percentage.

WP
A scalar representing the wilting point in percentage.

FC
A scalar representing the field capacity in percentage.

Value

A data frame with values of relative water content (RWC) and the effects of RWC on decomposition rates (fRWC).

References

Del Grosso, S. J., W. J. Parton, A. R. Mosier, E. A. Holland, E. Pendall, D. S. Schimel, and D. S. Ojima (2005), Modeling soil CO2 emissions from ecosystems, Biogeochemistry, 73(1), 71-91.

[Package SoilR version 1.1-36 Index]