fW.Demeter {SoilR}	R Documentation

Effects of moisture on decomposition rates according to the DEMETER model

Description

Calculates the effects of soil moisture on decomposition rates according to the DEMETER model.

Usage

fW.Demeter(M, Msat = 100)

Arguments

M
A scalar or vector containing values of soil moisture for which the effects on decomposition rates are calculated.

Msat
A scalar representing saturated soil moisture.

Value

A scalar or a vector containing the effects of moisture on decomposition rates (unitless).

References

Foley, J. A. (1995), An equilibrium model of the terrestrial carbon budget, Tellus B, 47(3), 310-319.

[Package SoilR version 1.2.103 Index]