fW.Gompertz {SoilR}	R Documentation

Effects of moisture on decomposition rates according to the Gompertz function

Description

Calculates the effects of water content on decomposition rates.

Usage

fW.Gompertz(theta, a = 0.824, b = 0.308)

Arguments

theta
A scalar or vector containing values of volumetric soil water content.

a
Empirical parameter

b
Empirical parameter

References

I. Janssens, S. Dore, D. Epron, H. Lankreijer, N. Buchmann, B. Longdoz, J. Brossaud, L. Montagnani. 2003. Climatic Influences on Seasonal and Spatial Differences in Soil CO2 Efflux. In Valentini, R. (Ed.) Fluxes of Carbon, Water and Energy of European Forests. pp 235-253. Springer.

[Package SoilR version 1.2.105 Index]