fW.Moyano {SoilR}	R Documentation

Effects of moisture on decomposition rates according to the function proposed by Moyano et al. (2013)

Description

Calculates the effects of water content on decomposition rates.

Usage

fW.Moyano(theta,
a=3.11,
b=2.42)

Arguments

theta
A scalar or vector containing values of volumetric soil water content.

a
Empirical parameter

b
Empirical parameter

References

F. E. Moyano, S. Manzoni, C. Chenu. 2013 Responses of soil heterotrophic respiration to moisture availability: An exploration of processes and models. Soil Biology and Biochemistry, Volume 59, April 2013, Pages 72-85

[Package SoilR version 1.1-50 Index]