fW.RothC {SoilR}	R Documentation

Effects of moisture on decomposition rates according to the RothC model

Description

Calculates the effects of moisture (precipitation and pan evaporation) on decomposition rates according to the RothC model.

Usage

fW.RothC(P,
E,
S.Thick=23,
pClay=23.4,
pE=0.75,
bare=FALSE)

Arguments

P
A vector with monthly precipitation (mm).

E
A vector with same length with open pan evaporation or evapotranspiration (mm).

S.Thick
Soil thickness in cm. Default for Rothamsted is 23 cm.

pClay
Percent clay.

pE
Evaporation coefficient. If open pan evaporation is used pE=0.75. If Potential evaporation is used, pE=1.0.

bare
Logical. Under bare soil conditions, bare=TRUE. Dafault is set under vegetated soil.

Value

A data.frame with accumulated top soil moisture deficit (Acc.TSMD) and the rate modifying factor b.

References

Coleman, K., and D. S. Jenkinson (1999), RothC-26.3 A model for the turnover of carbon in soil: model description and windows user guide (modified 2008), 47 pp, IACR Rothamsted, Harpenden.

[Package SoilR version 1.1-51 Index]