fW.Skopp {SoilR}	R Documentation

Effects of moisture on decomposition rates according to the function proposed by Skopp et al. 1990

Description

Calculates the effects of relative soil water content on decomposition rates.

Usage

fW.Skopp(rwc,
alpha=2,
beta=2,
f=1.3,
g=0.8)

Arguments

rwc
relative water content

alpha
Empirical parameter

beta
Empirical parameter

f
Empirical parameter

g
Empirical parameter

References

J. Skopp, M. D. Jawson, and J. W. Doran. 1990. Steady-state aerobic microbial activity as a function of soil water content. Soil Sci. Soc. Am. J., 54(6):1619-1625

[Package SoilR version 1.1-41 Index]