getAccumulatedRelease {SoilR}	R Documentation

getAccumulatedRelease S4 generic

Description

no Description

Usage

getAccumulatedRelease(object)

Arguments

object
A Model object (e.g. of class Model or Model14) Have a look at the methods for details.

Value

A n x m matrix of cummulative release fluxes with m columns representing the number of pools, and n rows representing the times as specified by the argument t in GeneralModel or other model creating functions.

Methods

getAccumulatedRelease,Model-method

[Package SoilR version 1.1-50 Index]