getAccumulatedRelease,Model-method {SoilR}	R Documentation

getAccumulatedRelease,Model-method time integrals of release fluxes per pool

Description

The method integrates the release flux of every pool for all times in the interval specified by the model definition.

Usage

## S4 method for signature 'Model'
getAccumulatedRelease(object)

Arguments

object
: of class Model

[Package SoilR version 1.1-50 Index]