getC {SoilR}	R Documentation

Calculates the content of the pools

Description

This function computes the content of the pools as function of time. In the original (and most of the present) Models these are Carbon pools hence the name. Have a look at the methods for details.

Usage

getC(object, as.closures = F)

Arguments

object
A modelrun as produced by the constructors: Model, Model_by_PoolNames, Model_14 the function GeneralModel or the functions listed in predefinedModels.
A model represents the initial value problem (IVP) for the contents of the pool consisting of

The initial values of the pool content

The system of ordinary differential equations, as dictated by the fluxes

The times for which the solution of the IVP is evaluated.

as.closures
see method arguments

Value

A matrix with m columns representing where m is the number of pools, and n rows where n is the number times as specified by the times of the model.

[Package SoilR version 1.2.104 Index]