getC,NlModel-method {SoilR}	R Documentation

Pool Contents for all times

Description

Pool Contents for all times

The solution of the initial value problem (IVP) for the pool contents. Since the first models in SoilR had only Carbon pools the function name getC could be interpreted as referring to the C content. If the model includes other element cycles e.g. N or P this interpretation is no longer valid. In this case the C in 'getC' stands for 'content' since the function will always return the solution for all pools, regardless of the chemical element the author of the model associated them with.

Usage

## S4 method for signature 'NlModel'
getC(object, as.closures = FALSE)

Arguments

object
object of class:NlModel, A modelrun as produced by the constructors: Model, Model_by_PoolNames, Model_14 the function GeneralModel or the functions listed in predefinedModels.
A model represents the initial value problem (IVP) for the contents of the pool consisting of

The initial values of the pool content

The system of ordinary differential equations, as dictated by the fluxes

The times for which the solution of the IVP is evaluated.

as.closures
If TRUE will return the result as a list of approximating functions of time indexed by the pool number.

Value

If as.closures is FALSE (the default) the return value is a matrix with as many columns as there are pools and as many rows as there are entries in the times argument the model has been built with.

[Package SoilR version 1.2.104 Index]