getC,Model-method {SoilR}	R Documentation

getC,Model-method

Description

This function computes the value for C for each time and pool.

Usage

## S4 method for signature 'Model'
getC(object)

Arguments

object
: of class Model

Value

A matrix with m columns representing the number of pools, and n rows representing the times as specified by the argument t in Model,GeneralModel or another model creating function.

[Package SoilR version 1.1-36 Index]