getC,NlModel-method {SoilR}	R Documentation

getC,NlModel-method compute the solution for the content of the pools

Description

This function computes the value for C for each time and pool.

Usage

## S4 method for signature 'NlModel'
getC(object,
as.closures=FALSE)

Arguments

object
: of class NlModel, the model

as.closures
: a flag that in TRUE will return an approximating function instead of the values

[Package SoilR version 1.1-41 Index]