getCompartmentalMatrixFunc,UnBoundNonLinDecompOp_by_PoolNames,character,character-method {SoilR}	R Documentation

Compartmental Matrix as function of the state vector and time

Description

Compartmental Matrix as function of the state vector and time

Usage

## S4 method for signature 
## 'UnBoundNonLinDecompOp_by_PoolNames,character,character'
getCompartmentalMatrixFunc(object, timeSymbol, state_variable_names)

Arguments

object
object of class:UnBoundNonLinDecompOp_by_PoolNames, An object of the class UnBoundNonLinDecompOp_by_PoolNames which is a representation equivalent to the compartmental matrix but independent of the order of state variables (pools) which therefore can be translated to any such ordering.

timeSymbol
object of class:character, The name of the argument representing time in the functions defining the fluxes in object

state_variable_names
object of class:character, The vector of the names of the state variables. The argument object is a representation of the compartmental system as #' lists of fluxes (internal fluxes and out-fluxes) as functions of the state variables and time. This method translates it to a matrix based formulation specific to a given ordering of the state variables. It is assumed (and checked) that the names formal arguments of the flux functions in object are a subset of the names of state_variable_names The method is used internally to translate the more intuitive (and more general) flux based description to the matrix based description required by the ode solvers.

[Package SoilR version 1.2.104 Index]