getDecompOp,Model-method {SoilR}	R Documentation

Extract the Compartmental Operator

Description

The method is usually used internally by other methods operating on models. The information it yields has either been provided by the user in creating the modelrun or can be obtained by directly transforming the arguments that were used.

Usage

## S4 method for signature 'Model'
getDecompOp(object)

Arguments

object
object of class:Model, A modelrun as produced by the constructors: Model, Model_by_PoolNames, Model_14 the function GeneralModel or the functions listed in predefinedModels.
A model represents the initial value problem (IVP) for the contents of the pool consisting of

The initial values of the pool content

The system of ordinary differential equations, as dictated by the fluxes

The times for which the solution of the IVP is evaluated.

Value

The actual class of the result can vary. It will be a subclass of DecompOp. These objects are an abstraction for a complete description of the fluxes in the pool system regardless of the form it is provided in. The information contained in these objects is equivalent to the set of internal and outward fluxes as functions of pool contents and time and sufficient to infer the "Compartmental Matrix" as a matrix valued function of the same arguments. In the general case of a nonautonomous nonlinear Model this function is a true function of both, the pool contents and time. In the case of an non-autonomous linear model it is a function of time only, and in case of a autonomous linear model it is a constant matrix. The vector valued function can be inferred by the generic function getFunctionDefinition.

[Package SoilR version 1.2.104 Index]