getF14C {SoilR}	R Documentation

Generic that yields the ^14C fraction for the cumulative content of all pools and all times

Description

Generic that yields the ^14C fraction for the cumulative content of all pools and all times

Usage

getF14C(object)

Arguments

object
see method arguments

[Package SoilR version 1.2-98 Index]