getF14R,Model_14-method {SoilR}	R Documentation

getF14R,Model_14-method average radiocarbon fraction weighted by carbonrelease

Description

Calculates the average radiocarbon fraction weighted by the amount of carbon release at each time step. (14R_1(t)+...+14R_n(t)) )/(R_1(t)+...R_n(t))) Where 14R_i(t) is the time dependent release of 14C of pool i and R_i(t) the release of all carbon isotops of pool i. Since the result is always in Absolute Fraction Modern format wie have to convert it to Delta14C

Usage

## S4 method for signature 'Model_14'
getF14R(object)

Arguments

object
: of class Model_14, an object

[Package SoilR version 1.1-50 Index]