getFunctionDefinition,UnBoundInFlux-method {SoilR}	R Documentation

getFunctionDefinition,UnBoundInFlux-method creates a constant timedependent function and returns it

Description

The method creates a timedependent function from the existing matrix describing the operator

Usage

## S4 method for signature 'UnBoundInFlux'
getFunctionDefinition(object= )

Arguments

object
: of class UnBoundInFlux

[Package SoilR version 1.1-49 Index]