getMeanTransitTime,ConstLinDecompOp-method {SoilR}	R Documentation

getMeanTransitTime,ConstLinDecompOp-method compute the mean transit time

Description

This method computes the mean transit time for the linear time invariant system that can be constructed from the given operator and input distribution.

It relies on the mehtod getTransitTimeDistributionDensity using the same arguments.

Usage

## S4 method for signature 'ConstLinDecompOp'
getMeanTransitTime(object,
inputDistribution)

Arguments

object
: of class ConstLinDecompOp

inputDistribution
: distribution of the inputs

Details

To compute the mean transit time for the distribution we have to compute the integral

\bar{T} = \int_0^∞ T \cdot S_r≤ft( \frac{\vec{I}}{I},0,T\right)\; dT

for the numerically computed density. To avoid issues with numerical integration we dont use ∞ as upper limit but cut off the integragion interval prematurely. For this purpose we calculate a maximum response time of the system as Lasaga

τ_{cycle} = \frac{1}{|\min(λ_i)|}

where λ_i are non-zero eigenvalues of the matrix {\bf A}.

References

Lasaga, A.: The kinetic treatment of geochemical cycles, Geochimica et Cosmochimica Acta, 44, 815 – 828, doi10.1016/0016-7037(80)90263-X, 1980.

[Package SoilR version 1.1-36 Index]