getOutputFluxes {SoilR}	R Documentation

Generic Function to obtain the fluxes out of of the pools

Description

Generic Function to obtain the fluxes out of of the pools

Usage

getOutputFluxes(object, as.closures = F)

Arguments

object
see method arguments

as.closures
see method arguments

[Package SoilR version 1.2.105 Index]