getReleaseFlux,Model-method {SoilR}	R Documentation

The release fluxes [content]/[time] for all pools.

Description

The release fluxes [content]/[time] for all pools.

Usage

## S4 method for signature 'Model'
getReleaseFlux(object)

Arguments

object
object of class:Model, A modelrun as produced by the constructors: Model, Model_by_PoolNames, Model_14 the function GeneralModel or the functions listed in predefinedModels.
A model represents the initial value problem (IVP) for the contents of the pool consisting of

The initial values of the pool content

The system of ordinary differential equations, as dictated by the fluxes

The times for which the solution of the IVP is evaluated.

Value

A matrix with as many columns as there are pools and as many rows as there are entries in the times argument the model has been build with.

[Package SoilR version 1.2.105 Index]