getReleaseFlux,Model_by_PoolNames-method {SoilR}	R Documentation

automatic title

Description

automatic title

Usage

## S4 method for signature 'Model_by_PoolNames'
getReleaseFlux(object)

Arguments

object
object of class:Model_by_PoolNames, A model(run) as produced by the constructors

link{Model}

link{Model_by_PoolNames}

link{Model_14}

the function GeneralModel or the functions in predefinedModels A model represents the initial value problem (IVP) for the contents of the pool consisting of

The initial values of the pool content

The system of ordinary differential oqations, as dictatet by the fluxes

The times for which the IVP is solved

[Package SoilR version 1.1-93 Index]