getReleaseFlux,Model-method {SoilR}	R Documentation

getReleaseFlux,Model-method

Description

The method computes the release of carbon per time for all points in time specified in the Model objects time slot.

Usage

## S4 method for signature 'Model'
getReleaseFlux(object)

Arguments

object
: of class Model, an object of class Model created by a call to a constructor e.g. Model, GeneralModelor other model creating functions.

Value

A n x m matrix of release fluxes with m columns representing the number of pools, and n rows representing the values for each time step as specified by the argument t in Model, GeneralModel or another model creating function.

[Package SoilR version 1.1-36 Index]