getReleaseFlux,NlModel-method {SoilR}	R Documentation

getReleaseFlux,NlModel-method get the release rate for all pools

Description

The method computes the release of carbon per time for all points in time specified in the objects time slot.

Usage

## S4 method for signature 'NlModel'
getReleaseFlux(object)

Arguments

object
: of class NlModel, an object of class NlModel

[Package SoilR version 1.1-36 Index]