getReleaseFlux14,Model_14-method {SoilR}	R Documentation

getReleaseFlux14,Model_14-method 14C respiration rate for all pools

Description

The function computes the 14C release flux ( mass per time ) for all pools. Note that the respiration coefficients for 14C do not change in comparison to the total C case. The fraction of 14C lost by respiration is not greater for 14C although the decay is faster due to the contribution of radioactivity.

Usage

## S4 method for signature 'Model_14'
getReleaseFlux14(object)

Arguments

object
: of class Model_14, the model.

[Package SoilR version 1.1-40 Index]