getRightHandSideOfODE,Model-method {SoilR}	R Documentation

Derivative of the state variables as function

Description

For non-linear models or models with state dependent influxes the returned function is a true function of state and time For linear models with state independent influxes the returned function is in fact a function of time only.

Usage

## S4 method for signature 'Model'
getRightHandSideOfODE(object)

Arguments

object
object of class:Model, no manual documentation

Value

A function f(t)

[Package SoilR version 1.2.105 Index]