getSolution,Model_by_PoolNames-method {SoilR}	R Documentation

All Fluxes and stocks for all times

Description

All Fluxes and stocks for all times

Usage

## S4 method for signature 'Model_by_PoolNames'
getSolution(object, params)

Arguments

object
object of class:Model_by_PoolNames, A modelrun as produced by the constructors: Model, Model_by_PoolNames, Model_14 the function GeneralModel or the functions listed in predefinedModels.
A model represents the initial value problem (IVP) for the contents of the pool consisting of

The initial values of the pool content

The system of ordinary differential equations, as dictated by the fluxes

The times for which the solution of the IVP is evaluated.

params
no manual documentation

Value

A matrix with as many columns as there are pools and as many rows as there are entries in the times argument the model has been build with.

[Package SoilR version 1.2.103 Index]