getTimeRange,UnBoundLinDecompOp-method {SoilR}	R Documentation

Extracts the time interval for which the function is valid.

Description

Extracts the time interval for which the function is valid.

Usage

## S4 method for signature 'UnBoundLinDecompOp'
getTimeRange(object)

Arguments

object
object of class:UnBoundLinDecompOp, no manual documentation

[Package SoilR version 1.2.105 Index]