getTimeRange,UnBoundInFlux-method {SoilR}	R Documentation

getTimeRange,UnBoundInFlux-method return an (infinite) time range since the function is assumed to be valid for all times

Description

some functions dealing with DecompOps in general rely on this so we have to implement it even though the timerange is always the same: (-inf,inf)

Usage

## S4 method for signature 'UnBoundInFlux'
getTimeRange(object)

Arguments

object
: of class UnBoundInFlux

[Package SoilR version 1.1-46 Index]