getTimeRange,TimeMap-method {SoilR}	R Documentation

getTimeRange,TimeMap-method

Description

The method returns the time range of the given object It is probably mostly used internally to make sure that time dependent functions retrieved from data are not used outside the interval where they are valid.

Usage

## S4 method for signature 'TimeMap'
getTimeRange(object)

Arguments

object
: of class TimeMap, An object of class TimeMap or one that inherits from TimeMap

[Package SoilR version 1.1-36 Index]