getTimes,Model-method {SoilR}	R Documentation

Extract the times vector

Description

Since the times had to be provided to create the model this method yields no new information. It is usually called internally by other functions that deal with models.

Usage

## S4 method for signature 'Model'
getTimes(object)

Arguments

object
object of class:Model, A modelrun as produced by the constructors: Model, Model_by_PoolNames, Model_14 the function GeneralModel or the functions listed in predefinedModels.
A model represents the initial value problem (IVP) for the contents of the pool consisting of

The initial values of the pool content

The system of ordinary differential equations, as dictated by the fluxes

The times for which the solution of the IVP is evaluated.

[Package SoilR version 1.2-98 Index]