getTransitTimeDistributionDensity,ConstLinDecompOp-method {SoilR}	R Documentation

getTransitTimeDistributionDensity,ConstLinDecompOp-method

Description

This mehtod computes the probability density of the transit time of the linear time invariant system that can be constructed from the given operator and input distribution.

Usage

## S4 method for signature 'ConstLinDecompOp'
getTransitTimeDistributionDensity(object,
inputDistribution,
times)

Arguments

object
: of class ConstLinDecompOp

inputDistribution
: distribution of the inputs

times
: the points in time where the solution is sought

References

Manzoni, S., Katul, G.~G., and Porporato, A.: Analysis of soil carbon transit times and age distributions using network theories, J. Geophys. Res., 114,

[Package SoilR version 1.1-36 Index]