initialize,Model_14-method {SoilR}	R Documentation

Internal method to supervise creation of objects of this class

Description

It is usually not necessary for user code to call this method. It's purpose is to enforce some sanity checks since it gets automatically called by new whenever an object of this class is created

Usage

## S4 method for signature 'Model_14'
initialize(
  .Object,
  times = c(0, 1),
  mat = ConstLinDecompOp(matrix(nrow = 1, ncol = 1, 0)),
  initialValues = numeric(),
  initialValF = ConstFc(values = c(0), format = "Delta14C"),
  inputFluxes = BoundInFluxes(function(t) {     return(matrix(nrow = 1, ncol = 1, 1)) },
    0, 1),
  c14Fraction = BoundFc(function(t) {     return(matrix(nrow = 1, ncol = 1, 1)) }, 0, 1),
  c14DecayRate = 0,
  solverfunc = deSolve.lsoda.wrapper,
  pass = FALSE
)

Arguments

.Object
object of class:Model_14, no manual documentation

times
no manual documentation

mat
no manual documentation

initialValues
no manual documentation

initialValF
no manual documentation

inputFluxes
no manual documentation

c14Fraction
no manual documentation

c14DecayRate
no manual documentation

solverfunc
no manual documentation

pass
no manual documentation

[Package SoilR version 1.2.104 Index]