linesCPool {SoilR}	R Documentation

Add lines with the output of getC14, getC, or getReleaseFlux to an existing plot

Description

This function adds lines to a plot with the C content, the C release, or Delta 14C value of each pool over time. Needs as input a matrix obtained after a call to getC14, getC, or getReleaseFlux.

Usage

linesCPool(t,
mat,
col,
...)

Arguments

t
A vector containing the time points for plotting.

mat
A matrix object obtained after a call to getC14, getC, or getReleaseFlux.

col
A color palette specifying color lines for each pool (columns of mat).

...
Other arguments passed to plot.

[Package SoilR version 1.1-51 Index]