listProduct {SoilR}	R Documentation

tensor product of lists

Description

Creates a list of all combinations of the elements of the inputlists (like a "tensor product list " The list elements can be of any class. The function is used in examples and tests to produce all possible combinations of arguments to a function. look at the tests for example usage

Usage

listProduct(...)

Arguments

...
lists

Value

a list of lists each containing one combinations of the elements of the input lists

Examples

listProduct(list('a','b'),list(1,2))

[Package SoilR version 1.2.104 Index]