pathEntropy {SoilR}	R Documentation

Path Entropy

Description

Computes the entropy of particles passing through the whole network of compartments for a model at equilibrium

Usage

pathEntropy(A, u)

Arguments

A
A constant compartmental square matrix with cycling rates in the diagonal and transfer rates in the off-diagonal.

u
A one-column matrix defining the amount of inputs per compartment.

Value

A scalar value with the path entropy

References

Metzler, H. (2020). Compartmental systems as Markov chains : age, transit time, and entropy (T. Oertel-Jaeger, I. Pavlyukevich, and C. Sierra, Eds.) [PhD thesis](https://suche.thulb.uni-jena.de/Record/1726091651)

Examples

B6=matrix(c(-1,1,0,0,-1,1,0,0,-1),3,3); u6=matrix(c(1,0,0))
pathEntropy(A=B6, u=u6)

[Package SoilR version 1.2.105 Index]