plot,Model-method {SoilR}	R Documentation

Create an overview plot

Description

The method solves the model and plots the solutions It is intended to provide a quick overview.

Usage

## S4 method for signature 'Model'
plot(x)

Arguments

x
object of class:Model, The model (run) the results of which are plotted

[Package SoilR version 1.2.103 Index]