plotC14Pool {SoilR}	R Documentation

Plots the output of getF14 for each pool over time

Description

This function produces a plot with the Delta14C in the atmosphere and the Delta14C of each pool obtained after a call to getF14.

Usage

plotC14Pool(t, mat, inputFc, col, ...)

Arguments

t
A vector containing the time points for plotting.

mat
A matrix object obtained after a call to getF14

inputFc
A Data Frame object containing values of atmospheric Delta14C per time. First column must be time values, second column must be Delta14C values in per mil.

col
A color palette specifying color lines for each pool (columns of mat).

...
Other arguments passed to plot.

[Package SoilR version 1.2.104 Index]