plotCPool {SoilR}	R Documentation

Plots the output of getC or getReleaseFlux for each pool over time

Description

This function produces a plot with the C content or released C for each pool over time. Needs as input a matrix obtained after a call to getC or getReleaseFlux.

Usage

plotCPool(t,
mat,
col,
...)

Arguments

t
A vector containing the time points for plotting.

mat
A matrix object obtained after a call to getC or getReleaseFlux

col
A color palette specifying color lines for each pool (columns of mat).

...
Other arguments passed to link{plot}.

[Package SoilR version 1.1-53 Index]