plotPoolGraph {SoilR}	R Documentation

Generic plotter

Description

Generic plotter

Usage

plotPoolGraph(x)

Arguments

x
An argument containing sufficient information about the connections between the pools as well as from and to the exterior.

[Package SoilR version 1.2.103 Index]