[,Model,character,missing,missing-method {SoilR}	R Documentation

Experimentally overloaded index operator

Description

The method provides shortcuts and a unified interface to some of the methods that can be applied to a model. For a given model ‘M' the code 'M[’C'] is equivalent to 'getC(M)' and ‘M[’ReleaseFlux']' is equivalent to 'getReleaseFlux(M)' ‘M[’AccumulatedRelease']' is equivalent to 'getAccumulatedRelease(M)'

Usage

## S4 method for signature 'Model,character,missing,missing'
x[i]

Arguments

x
object of class:Model, no manual documentation

i
object of class:character, no manual documentation

[Package SoilR version 1.1-93 Index]