transitTime {SoilR}	R Documentation

Transit times for compartment models

Description

Computes the density distribution and mean for the transit time of a compartmental model

Usage

transitTime(A,
u,
a=seq(0, 100),
q=c(0.05, 0.5, 0.95))

Arguments

A
A compartmental linear square matrix with cycling rates in the diagonal and transfer rates in the off-diagonal.

u
A one-column matrix defining the amount of inputs per compartment.

a
A sequence of ages to calculate density functions

q
Vector of probabilities to calculate quantiles of the transit time distribution

Value

A list with 3 objects: mean transit time, transit time density distribution, and quantiles.

See Also

systemAge

[Package SoilR version 1.1-53 Index]