turnoverFit {SoilR}	R Documentation

Estimation of the turnover time from a radiocarbon sample.

Description

This function finds two possible values of turnover time from radiocarbon sample assuming a one pool model with carbon at equilibrium.

Usage

turnoverFit(obsC14, obsyr, yr0, Fatm, plot = TRUE, by = 0.5)

Arguments

obsC14
a scalar with the observed radiocarbon value in Delta14C

obsyr
a scalar with the year in which the sample was taken.

yr0
The year at which simulations will start.

Fatm
an atmospheric radiocarbon curve as data.frame. First column must be time.

plot
logical. Should the function produce a plot?

by
numeric. The increment of the sequence of years used in the simulations.

Details

This algorithm takes an observed radiocarbon value and runs OnepModel14, calculates the squared difference between predictions and observations, and uses optimize to find the minimum difference.

Value

A numeric vector with two values of the turnover time that minimize the difference between the prediction of a one pool model and the observed radiocarbon value.

[Package SoilR version 1.2.105 Index]