$,NlModel-method {SoilR}	R Documentation

$,NlModel-method overload the $ operator

Description

This function is still experimental and might change considerably It is meant as an alternative to all the get methods to provide a more R like interface.

Usage

## S4 method for signature 'NlModel'
x$
name

Arguments

x
: of class NlModel, the model

name
: the property to access

[Package SoilR version 1.1-49 Index]