[,NlModel,character,ANY,ANY-method {SoilR}	R Documentation

[,NlModel,character,ANY,ANY-method overloads the [] operator

Description

This function is still experimental and might change considerably It is meant as an alternative to all the get methods to provide a more R like interface.

Usage

## S4 method for signature 'NlModel,character,ANY,ANY'
x[
i]

Arguments

x
: of class NlModel, the model

i
: of class character, the property to access

[Package SoilR version 1.1-40 Index]