BoundInFlux-class {SoilR}	R Documentation

BoundInFlux S4 class

Description

defines a time dependent inputrate as function of time and including the domain where the function is well defined. This can be used to avoid interpolations out of range when mixing different time dependent data sets

Methods

No methods defined with class BoundInFlux in the signature.

Superclasses

InFlux-class
TimeMap-class

Constructors

BoundInFlux

[Package SoilR version 1.1-50 Index]