BoundInFlux-class {SoilR}	R Documentation

BoundInFlux S4 class

Description

defines a time dependent inputrate as function of time and including the domain where the function is well defined. This can be used to avoid interpolations out of range when mixing different time dependent data sets

Methods

No exported methods directly defined for class BoundInFlux:

Methods inherited from superclasses:

from class InFlux:

GeneralInFlux

signature(object = "InFlux"): ...

GeneralInFlux,InFlux-method from class TimeMap:

GeneralDecompOp

signature(object = "TimeMap"): ...

GeneralDecompOp,TimeMap-method

GeneralInFlux

signature(object = "TimeMap"): ...

GeneralInFlux,TimeMap-method

TimeMap

signature(map = "TimeMap", starttime = "ANY", endtime = "ANY", times = "ANY", data = "ANY"): ...

TimeMap,TimeMap,ANY,ANY,ANY,ANY-method

add_plot

signature(x = "TimeMap"): ...

add_plot,TimeMap-method

as.character

signature(x = "TimeMap"): ...

as.character,TimeMap-method

getFunctionDefinition

signature(object = "TimeMap"): ...

getFunctionDefinition,TimeMap-method

getTimeRange

signature(object = "TimeMap"): ...

getTimeRange,TimeMap-method

[Package SoilR version 1.1-82 Index]