CenturyModel {SoilR}	R Documentation

Implementation of the Century model

Description

This function implements the Century model as described in Parton et al. (1987).

Usage

CenturyModel(
  t,
  ks = c(k.STR = 0.094, k.MET = 0.35, k.ACT = 0.14, k.SLW = 0.0038, k.PAS = 0.00013),
  C0 = c(0, 0, 0, 0, 0),
  In,
  LN,
  Ls,
  clay = 0.2,
  silt = 0.45,
  xi = 1,
  xi_lag = 0,
  solver = deSolve.lsoda.wrapper
)

Arguments

t
A vector containing the points in time where the solution is sought.

ks
A vector of length 5 containing the values of the decomposition rates for the different pools. Units in per week.

C0
A vector of length 5 containing the initial amount of carbon for the 5 pools.

In
A scalar or data.frame object specifying the amount of litter inputs by time (mass per area per week).

LN
A scalar representing the lignin to nitrogen ratio of the plant residue inputs.

Ls
A scalar representing the fraction of structural material that is lignin.

clay
Proportion of clay in mineral soil.

silt
Proportion of silt in mineral soil.

xi
A scalar, data.frame, function or anything that can be converted to a scalar function of time ScalarTimeMap #' object specifying the external (environmental and/or edaphic) effects on decomposition rates.

xi_lag
A time shift/delay for the automatically created time dependent function xi(t)

solver
A function that solves the system of ODEs. This can be euler or deSolve.lsoda.wrapper or any other user provided function with the same interface.

Value

A Model Object that can be further queried

References

Parton, W.J, D.S. Schimel, C.V. Cole, and D.S. Ojima. 1987. Analysis of factors controlling soil organic matter levels in Great Plain grasslands. Soil Science Society of America Journal 51: 1173–1179. Sierra, C.A., M. Mueller, S.E. Trumbore. 2012. Models of soil organic matter decomposition: the SoilR package version 1.0. Geoscientific Model Development 5, 1045-1060.

See Also

RothCModel. There are other predefinedModels and also more general functions like Model.

[Package SoilR version 1.2.104 Index]