ConstantOutFluxRateList_by_PoolName-class {SoilR}	R Documentation

Subclass of list that is guaranteed to contain only elements of type ConstantOutFluxRate_by_PoolName

Description

The main purpose of the class is to be used in method signatures which would otherwise only indicate an object of class 'list' in their signature an then check that the list contains the right kind of elements inside the function. Using this class the method signature becomes much more informative.

[Package SoilR version 1.2.104 Index]