GeneralModel_14 {SoilR}	R Documentation

a function to create objects of class Model_14

Description

At the moment this is just a wrapper for the actual constructor Model_14 with additional support for some now deprecated parameters for backward compatibility. This role may change in the future to an abstract factory where the actual class of the created model will be determined by the supplied parameters.

Usage

GeneralModel_14(t= ,
A= ,
ivList= ,
initialValF= ,
inputFluxes= ,
Fc=NULL,
inputFc=Fc,
di=-0.0001209681,
solverfunc=deSolve.lsoda.wrapper,
pass=FALSE)

Arguments

t

A

ivList

initialValF

inputFluxes

Fc
deprecated keyword argument, please use inputFc instead

inputFc

di

solverfunc

pass

Value

A model object that can be further queried.

See Also

TwopParallelModel, TwopSeriesModel, TwopFeedbackModel

[Package SoilR version 1.1-49 Index]