GeneralModel_14 {SoilR}	R Documentation

a function to create objects of class Model_14

Description

This method tries to create a Model object from any combination of arguments that can be converted into the required set of building blocks for a model for n arbitrarily connected pools.

Usage

GeneralModel_14(t,
A,
ivList,
initialValF,
inputFluxes,
Fc,
di=-0.0001209681,
solverfunc=deSolve.lsoda.wrapper,
pass=FALSE)

Arguments

t
A vector containing the points in time where the solution is sought.

A

ivList

initialValF

inputFluxes

Fc

di

solverfunc

pass

Value

A model object that can be further queried.

See Also

TwopParallelModel, TwopSeriesModel, TwopFeedbackModel

[Package SoilR version 1.1-46 Index]