GeneralModel_14 {SoilR}	R Documentation

GeneralModel_14 S4 generic

Description

no Description

Usage

GeneralModel_14(t,
A,
ivList,
initialValF,
inputFluxes,
inputFc,
Fc,
di=-0.0001209681,
lambda=-0.0001209681,
solverfunc=deSolve.lsoda.wrapper,
pass=FALSE,
...)

Arguments

t
see the method arguments for details

A
see the method arguments for details

ivList
see the method arguments for details

initialValF
see the method arguments for details

inputFluxes
see the method arguments for details

inputFc
see the method arguments for details

Fc
see the method arguments for details

di
see the method arguments for details

lambda
see the method arguments for details

solverfunc
The function used by to actually solve the ODE system. This can be deSolve.lsoda.wrapper or any other user provided function with the same interface.

pass
if TRUE Forces the constructor to create the model even if it is invalid

...
see the method arguments for details

Methods

GeneralModel_14,numeric,ANY,numeric,ANY,ANY,ANY,missing,missing,missing,missing,logical-method
GeneralModel_14,numeric,ANY,numeric,ANY,ANY,ANY,missing,numeric,missing,function,logical-method
GeneralModel_14,numeric,ANY,numeric,ANY,ANY,ANY,missing,numeric,missing,function,missing-method
GeneralModel_14,numeric,ANY,numeric,ANY,ANY,ANY,missing,numeric,missing,missing,logical-method
GeneralModel_14,numeric,ANY,numeric,ANY,ANY,ANY,missing,numeric,missing,missing,missing-method
GeneralModel_14,numeric,ANY,numeric,ANY,ANY,missing,ANY,numeric,missing,missing,missing-method

[Package SoilR version 1.1-40 Index]