InFlux-class {SoilR}	R Documentation

InFlux S4 class

Description

All models need to specify the influx of material to the pools. This parameter will be represented as an object of one of the subclasses of this class. The most general form of influx supported up to now is a vector valued function of time represented by BoundInFlux-class. In the most simple case it is constant and represented by an object of class ConstInFlux-class. Such an object can for instance be created from a numeric vector.

Methods

Exported methods directly defined for class InFlux:

GeneralInFlux

signature(object = "InFlux"): ...

GeneralInFlux,InFlux-method

[Package SoilR version 1.1-82 Index]