InFluxList_by_PoolName-class {SoilR}	R Documentation

Class for a list of influxes indexed by the names of the target pools

Description

Class for a list of influxes indexed by the names of the target pools

S4-methods

S4-methods with class InFluxList_by_PoolName in their signature:

as.numeric,InFluxList_by_PoolName-method

by_PoolIndex,InFluxList_by_PoolName,character,character-method

getFunctionDefinition,InFluxList_by_PoolName-method

S4-methods with superclasses (in the package) of class InFluxList_by_PoolName in their signature:

superclass list

ConstantInFluxList_by_PoolIndex,list-method

ConstantInternalFluxRateList_by_PoolIndex,list-method

ConstantInternalFluxRateList_by_PoolName,list-method

ConstantOutFluxRateList_by_PoolIndex,list-method

ConstantOutFluxRateList_by_PoolName,list-method

GeneralDecompOp,list-method

InFluxes,list-method

InFluxList_by_PoolIndex,list-method

InFluxList_by_PoolName,list-method

InternalFluxList_by_PoolIndex,list-method

InternalFluxList_by_PoolName,list-method

OutFluxList_by_PoolIndex,list-method

OutFluxList_by_PoolName,list-method

StateIndependentInFluxList_by_PoolIndex,list-method

TimeMap,list,missing,missing,missing,missing-method

TimeMap,missing,missing,missing,numeric,list-method

[Package SoilR version 1.2-98 Index]