Model_14-class {SoilR}	R Documentation

Model_14

Description

This class extends Model, to represent 14C decay.

As Model it contains all the components that are needed to solve the initial value problem for the pool contents. C=(C_1,...C_n)^t, .

It adds the components that are needed to solve the additional initial value problem for the 14C contents of the pools. 14C=(14C_1,...14C_n)^t, .

It provides the single argument for all the different functions that are available for arguments of class Model.

and for additional functions that are available to compute various results from the solution of the additional initial value problem for 14C. See subsection Methods and the examples.)

Details

The additional initial value problem is given by:

the ordinary differential equation d/dt 14C=(A(t)+kI) 14C+I(t),

the initial Values C_0=C(t_0),

the 14C fraction

the 14C decay rate k.

for the times {t_0,....,t_m}.

In an object of this class the components are represented as follows:

The time-dependent matrix valued function A(t) is represented by an object of a class that inherits from class DecompOp. Such objects can be created in different ways from functions, matrices or data. (see the subclasses of DecompOp and especially their Constructors sections. and the examples section of this help page.

The vector-valued time-dependent function I(t) is in SoilR represented by an object of a class that inherits from class InFlux InFlux. Such objects can be created from functions, constant vectors and data. (see the subclasses of InFlux and especially their Constructors sections.

The times for which the results are computed are represented by a numeric vector.

The initial values are represented by a numeric vector

Methods

getC14

signature(object = "Model_14"): ...

getC14,Model_14-method

getF14C

signature(object = "Model_14"): ...

getF14C,Model_14-method

getF14R

signature(object = "Model_14"): ...

getF14R,Model_14-method

getF14

signature(object = "Model_14"): ...

getF14,Model_14-method

getReleaseFlux14

signature(object = "Model_14"): ...

getReleaseFlux14,Model_14-method

Exported superclasses

Model-class

Constructors

Model_14

[Package SoilR version 1.1-46 Index]