PoolName-class {SoilR}	R Documentation

class for pool-name-strings

Description

used to control the creation of PoolName objects which have to be valid R identifiers and to dispatch pool name specific methods like conversion to pool indices

S4-methods

S4-methods with class PoolName in their signature:

PoolIndex,PoolName-method

PoolName,PoolName-method

S4-methods with superclasses (in the package) of class PoolName in their signature:

superclass character

[,Model,character,missing,missing-method

[,NlModel,character,ANY,ANY-method

BoundFc,character-method

by_PoolIndex,function,character,character-method

by_PoolIndex,InFlux_by_PoolName,character,character-method

by_PoolIndex,InFluxList_by_PoolName,character,character-method

by_PoolIndex,InternalFlux_by_PoolName,character,character-method

by_PoolIndex,InternalFluxList_by_PoolName,character,character-method

by_PoolIndex,OutFlux_by_PoolName,character,character-method

by_PoolIndex,OutFluxList_by_PoolName,character,character-method

ConstantInternalFluxRate_by_PoolIndex,missing,missing,character,numeric-method

ConstantInternalFluxRate_by_PoolName,character,character,missing,numeric-method

ConstantInternalFluxRate_by_PoolName,missing,missing,character,numeric-method

ConstLinDecompOp,missing,ConstantInternalFluxRateList_by_PoolName,ConstantOutFluxRateList_by_PoolName,missing,character-method

GeneralPoolId,character-method

getCompartmentalMatrixFunc,UnBoundNonLinDecompOp_by_PoolNames,character,character-method

InFlux_by_PoolName,function,character-method

InternalFlux_by_PoolName,function,character,character,missing-method

InternalFlux_by_PoolName,function,missing,missing,character-method

OutFlux_by_PoolName,function,character-method

PoolConnection_by_PoolIndex,missing,missing,character-method

PoolIndex,character-method

PoolName,character-method

UnBoundNonLinDecompOp_by_PoolNames,InternalFluxList_by_PoolName,OutFluxList_by_PoolName,character-method

UnBoundNonLinDecompOp,missing,missing,missing,missing,character,character,UnBoundNonLinDecompOp_by_PoolNames-method

S4-superclasses (in the package)

PoolId

character

[Package SoilR version 1.2.104 Index]