RespirationCoefficients {SoilR}	R Documentation

helper function to compute respiration coefficients

Description

This function computes the respiration coefficients as function of time for all pools according to the given matrix function A(t)

Usage

RespirationCoefficients(A)

Arguments

A
A matrix valued function representing the model.

Value

A vector valued function of time containing the respiration coefficients for all pools.

[Package SoilR version 1.1-97 Index]