ScalarTimeMap-class {SoilR}	R Documentation

S4 class for a scalar time dependent function on a finite time interval

Description

S4 class for a scalar time dependent function on a finite time interval

S4-methods

S4-methods with class ScalarTimeMap in their signature:

ConstLinDecompOpWithLinearScalarFactor,matrix,missing,missing,missing,ScalarTimeMap-method

S4-methods with superclasses (in the package) of class ScalarTimeMap in their signature:

superclass TimeMap

add_plot,TimeMap-method

as.character,TimeMap-method

GeneralDecompOp,TimeMap-method

getFunctionDefinition,TimeMap-method

getTimeRange,TimeMap-method

InFluxes,TimeMap-method

initialize,TimeMap-method

plot,TimeMap-method

TimeMap,TimeMap,ANY,ANY,ANY,ANY-method

S4-subclasses

BoundFc

S4-superclasses (in the package)

TimeMap

[Package SoilR version 1.2.105 Index]