TimeMap-class {SoilR}	R Documentation

TimeMap S4 class

Description

This class enhances a time dependent function by information about its domain and a possible delay. The information about the delay is especially usefull for functions that interpolate data. Assume that you are give time series data in two vectors times, values. You can create an interpolating function with splinefun or approxfun f <- splinefun(x=times,y=values) f(t) will yield sensible values for min(times)<t<max(times). but will produce unreasonable values for any t outside these limits. Unfortunately the interpolating functions produced by splinefun or approxfun do not retain any information about their domain which makes it possible to accidentatly apply them to times not at all supported by the original data. This would not even cause errors in the code but silently corrupt the results. To help you to keep track of the domains of the many time dependent functions used in SoilR's Models thss class TimeMap-class stores the starttime and endtime values along with the function represented by map. SoilR functions that accept time series data will normally convert it to subclasses TimeMap-class automatically but you can do it explicitly.

Methods

GeneralDecompOp

signature(object = "TimeMap"): ...

GeneralDecompOp,TimeMap-method

GeneralInFlux

signature(object = "TimeMap"): ...

GeneralInFlux,TimeMap-method

TimeMap

signature(map = "TimeMap", starttime = "ANY", endtime = "ANY", times = "ANY", data = "ANY"): ...

TimeMap,TimeMap,ANY,ANY,ANY,ANY-method

as.character

signature(x = "TimeMap"): ...

as.character,TimeMap-method

getFunctionDefinition

signature(object = "TimeMap"): ...

getFunctionDefinition,TimeMap-method

getTimeRange

signature(object = "TimeMap"): ...

getTimeRange,TimeMap-method

Subclasses

BoundInFlux
TransportDecompositionOperator
BoundLinDecompOp
BoundFc

Constructors

TimeMap
Please also look at constructors of non virtual subclasses: BoundInFlux, TransportDecompositionOperator, BoundLinDecompOp, BoundFc.

[Package SoilR version 1.1-49 Index]