TimeMap-class {SoilR}	R Documentation

S4 class for a time dependent function

Description

The class represents functions which are defined on a (possibly infinite) interval from [starttime,endtime] Instances are usually created internally from data frames or lists provided by the user in the high level interfaces.

Details

The class is necessary to be able to detect unwanted extrapolation of time line data which might otherwise occur for some of the following reasons: SoilR allows to specify measured data for many of its arguments and computes the interpolating functions automatically. The functions returned by the standard R interpolation mechanisms like splinefun or approxfun do not provide a safeguard against accidental extrapolation. Internally SoilR converts nearly all data to time dependent functions e.g. to be used in ode solvers. So the information of the domain of the function has to be kept.

S4-methods

S4-methods with class TimeMap in their signature:

add_plot,TimeMap-method

as.character,TimeMap-method

GeneralDecompOp,TimeMap-method

getFunctionDefinition,TimeMap-method

getTimeRange,TimeMap-method

InFluxes,TimeMap-method

initialize,TimeMap-method

plot,TimeMap-method

TimeMap,TimeMap,ANY,ANY,ANY,ANY-method

S4-subclasses

ScalarTimeMap

BoundLinDecompOp

TransportDecompositionOperator

BoundInFluxes

BoundFc

[Package SoilR version 1.1-93 Index]