TimeMap.new {SoilR}	R Documentation

deprecated constructor of the class TimeMap.

Description

deprecated functions #################### use the generic TimeMap(...) instead

Usage

TimeMap.new(t_start,
t_end,
f)

Arguments

t_start
A number marking the begin of the time domain where the function is valid

t_end
A number the end of the time domain where the function is valid

f
The time dependent function definition (a function in R's sense)

Value

An object of class TimeMap that can be used to describe models.

[Package SoilR version 1.1-40 Index]