fT.Daycent1 {SoilR}	R Documentation

Effects of temperature on decomposition rates according to the DAYCENT model

Description

Calculates the effects of temperature on decomposition rates according to the DAYCENT model.

Usage

fT.Daycent1(Temp)

Arguments

Temp
A scalar or vector containing values of soil temperature for which the effects on decomposition rates are calculated

Value

A scalar or a vector containing the effects of temperature on decomposition rates (unitless).

References

Kelly, R. H., W. J. Parton, M. D. Hartman, L. K. Stretch, D. S. Ojima, and D. S. Schimel (2000), Intra-annual and interannual variability of ecosystem processes in shortgrass steppe, J. Geophys. Res., 105.

[Package SoilR version 1.1-82 Index]