fT.LandT {SoilR}	R Documentation

Effects of temperature on decomposition rates according to a function proposed by Lloyd and Taylor (1994)

Description

Calculates the effects of temperature on decomposition rates according to a function proposed by Lloyd and Taylor (1994).

Usage

fT.LandT(Temp)

Arguments

Temp
A scalar or vector containing values of soil temperature for which the effects on decomposition rates are calculated

Value

A scalar or a vector containing the effects of temperature on decomposition rates (unitless).

References

Lloyd, J., and J. A. Taylor (1994), On the Temperature Dependence of Soil Respiration, Functional Ecology, 8(3), 315-323.

[Package SoilR version 1.1-53 Index]