getAccumulatedRelease,Model-method {SoilR}	R Documentation

automatic title

Description

automatic title

Usage

## S4 method for signature 'Model'
getAccumulatedRelease(object)

Arguments

object
object of class:Model, A model(run) as produced by the constructors

Model

Model_by_PoolNames

Model_14

the function GeneralModel or the functions listed in predefinedModels. A model represents the initial value problem (IVP) for the contents of the pool consisting of

The initial values of the pool content

The system of ordinary differential oqations, as dictated by the fluxes

The times for which the solution of the IVP is evaluated.

[Package SoilR version 1.1-93 Index]