getC {SoilR}	R Documentation

Generic Function to obtain the contents of the pools for all time steps

Description

Generic Function to obtain the contents of the pools for all time steps

Usage

getC(object, as.closures = F)

Arguments

object
see method arguments

as.closures
see method arguments

[Package SoilR version 1.1-93 Index]