getC {SoilR}	R Documentation

getC S4 generic

Description

no Description

Usage

getC(object,
as.closures=F)

Arguments

object
some model object, the actual class depends on the method used.

as.closures
if set to TRUE instead of a matrix a list of functions will be returned.

Value

A matrix with m columns representing the number of pools, and n rows representing the times as specified by the argument t in GeneralModel or another model creating function.

Methods

getC,Model-method
getC,NlModel-method

[Package SoilR version 1.1-47 Index]