getC,Model-method {SoilR}	R Documentation

getC,Model-method

Description

This function computes the value for C for each time and pool.

Usage

## S4 method for signature 'Model'
getC(object= )

Arguments

object
: of class Model

Details

This function takes a Model object, which represents a system of ODEs of the form

dC(t)/dt = I(t) + A(t)C(t)

and solves the system for C(t). The numerical solver used can be specified in the constructor of the Model class e.g. Model, GeneralModel.

Value

A matrix with m columns representing the number of pools, and n rows representing the times as specified by the argument t in Model,GeneralModel or another model creating function.

See Also

See examples in GeneralModel, GeneralModel_14, TwopParallelModel, TwopSeriesModel, TwopFeedbackModel, etc.

[Package SoilR version 1.1-49 Index]