getMeanTransitTime {SoilR}	R Documentation

This function

Description

no Description

Usage

getMeanTransitTime(object,
inputDistribution)

Arguments

object
a DecompOp Object.

inputDistribution
a vector of length equal to the number of pools. The entries are weights, which must sum to 1.

Methods

getMeanTransitTime,ConstLinDecompOp-method

References

Manzoni, S., G.G. Katul, and A. Porporato. 2009. Analysis of soil carbon transit times and age distributions using network theories. Journal of Geophysical Research-Biogeosciences 114, DOI: 10.1029/2009JG001070.

Thompson, M.~V. and Randerson, J.~T.: Impulse response functions of terrestrial carbon cycle models: method and application, Global Change Biology, 5, 371–394, 10.1046/j.1365-2486.1999.00235.x, 1999.

Bolin, B. and Rodhe, H.: A note on the concepts of age distribution and transit time in natural reservoirs, Tellus, 25, 58–62, 1973.

Eriksson, E.: Compartment Models and Reservoir Theory, Annual Review of Ecology and Systematics, 2, 67–84, 1971.

[Package SoilR version 1.1-49 Index]