getReleaseFlux,Model-method {SoilR}	R Documentation

getReleaseFlux,Model-method get the release rate for all pools

Description

The method computes the release of carbon per time for all points in time specified in the Model objects time slot.

Usage

## S4 method for signature 'Model'
getReleaseFlux(object)

Arguments

object
: of class Model, an object of class Model created by a call to a constructor e.g. Model, GeneralModelor other model creating functions.

Details

This function takes a Model object, which represents a system of ODEs

dC(t)/dt = I(t) + A(t)C(t)

solves the system for C(t), calculates the release coefficients R(t), and computes the release flux as R(t) C(t). The numerical solver used can be specified in the model creating functions like e.g. Model.

Value

A n x m matrix of release fluxes with m columns representing the number of pools, and n rows representing the values for each time step as specified by the argument t in Model, GeneralModel or another model creating function.

[Package SoilR version 1.1-40 Index]