getTransitTimeDistributionDensity {SoilR}	R Documentation

This function

Description

no Description

Usage

getTransitTimeDistributionDensity(object,
inputDistribution,
times)

Arguments

object
a protoDecompOp Object

inputDistribution
a vector of length equal to the number of pools. The entries are weights. That means that their sume must be equal to one!

times
the times for which the distribution density is sought

Methods

getTransitTimeDistributionDensity,ConstLinDecompOp-method

References

Manzoni, S., G.G. Katul, and A. Porporato. 2009. Analysis of soil carbon transit times and age distributions using network theories. Journal of Geophysical Research-Biogeosciences 114, DOI: 10.1029/2009JG001070.

[Package SoilR version 1.1-51 Index]