getTransitTimeDistributionDensity,ConstLinDecompOp-method {SoilR}	R Documentation

getTransitTimeDistributionDensity,ConstLinDecompOp-method compute the TransitTimeDistributionDensity

Description

This mehtod computes the probability density of the transit time of the linear time invariant system that can be constructed from the given operator and input distribution.

Usage

## S4 method for signature 'ConstLinDecompOp'
getTransitTimeDistributionDensity(object,
inputDistribution,
times)

Arguments

object
: of class ConstLinDecompOp

inputDistribution
: distribution of the inputs

times
: the points in time where the solution is sought

Details

In a forthcoming paper SoilRv1.2 we derive the algorithm used in this implementation under the assumption of steady conditions having prevailed infinitely. We arrive at a formulation well known from the literature about time invariant linlear systems, cited e.g. in ManzoniJGR.
The somehow amazing result is that the weight of the transit time density ψ(T) for a transit time T for the steady state system is identical to the output O(T) observed at time T of a different system which started with a normalized impulsive input \frac{\vec{I}}{I} at time T=0, where I=∑_{k=1}^m i_k is the cumulative input flux to all pools.
This fact simpliefies the computation considerably. Translated into the language of an ode solver an impulsive input becomes a start vector \frac{\vec{I}}{I} at time T=0 and O(T) the respiration related to the solution of the initial value problem observed at time T.

Note that from the perspective of the ode solver S_r depends on the decomposition operator and the distribution of the input among the pools only. It is therefor possible to implement the transit time distribution as a function of the decomposition operator and the fixed input flux distribution. To insure steady state conditions the decomposition operator is not allowed to be a true function of time. We therefor implement the method only for the subclass ConstLinDecompOp
Remark:
The decision to implement this method for transitTimeDensity especially for objects of class ConstLinDecompOp reflects the fact that an arbitrary model in SoilR is by no means bound to end up in steady state. To insure this we would have to ignore the input part of a user created model which would be confusing.
Remark:
In future versions of SoilR it will be possible to compute a dynamic, time dependent transit time distribution for objects of class Model with a time argument specifying for which time the distribution is sought. The steady state results computed here could than be reproduced with the user responsible for providing a model actually reaching it.

References

Manzoni, S., Katul, G.~G., and Porporato, A.: Analysis of soil carbon transit times and age distributions using network theories, J. Geophys. Res., 114,

[Package SoilR version 1.1-49 Index]