turnoverFit {SoilR}	R Documentation

Estimation of the turnover time from a soil radiocarbon sample.

Description

This function finds the best possible value of turnover time from a soil radiocarbon sample assuming a one pool model and annual litter inputs.

Usage

turnoverFit(
  obsC14,
  obsyr,
  In,
  C0 = 0,
  yr0 = 1900,
  Zone = "NHZone2",
  plot = TRUE,
  by = 0.5
)

Arguments

obsC14
a scalar with the observed radiocarbon value in Delta14C of the soil sample.

obsyr
a scalar with the year in which the soil sample was taken.

In
a scalar or data.frame with the annual amount of litter inputs to the soil.

C0
a scalar with the initial amount of carbon stored at the beginning of the simulation.

yr0
The year at which simulations will start.

Zone
the hemispheric zone of atmospheric radiocarbon. Possible values are NHZone1: northern hemisphere zone 1, NHZone2: northern hemisphere zone 2, NHZone3: northern hemisphere zone 3, SHZone12: southern hemisphere zones 1 and 2, SHZone3: southern hemisphere zone 3. See Hua2013 for additional information.

plot
logical. Should the function produce a plot?

by
numeric. The increment of the sequence of years used in the simulations.

Details

This algorithm takes the observed values and a given amount of litter inputs, runs OnepModel14, calculates the squared difference between predictions and observations, and uses optimize to find the minimum difference. If the turnover time is relatively short (< 50 yrs), it is safe to assume C0=0 because the soil will reach steady state within the simulation time. However, for longer turnover times it is recommended to use a value of C0 close to the steady state value.

Value

A scalar with the value of the turnover time that minimizes the difference between the prediction of a one pool model and the observed radiocarbon value.

[Package SoilR version 1.1-97 Index]